Crate k8s_openapi

Bindings for the Kubernetes client API, generated from the OpenAPI spec.

Each supported version of Kubernetes is represented by a feature name (like v1_9). Only one such feature can be enabled at a time.

These docs have been generated with the v1_14 feature enabled. To see docs for one of the other supported versions, please generate the docs locally with cargo doc --features 'v1_<>'

If you're writing an application crate, your crate must depend on k8s-openapi and enable the feature corresponding to the version of Kubernetes that your application supports.

# For application crates

[dependencies]
k8s-openapi = { version = "...", features = ["v1_14"] }

If you're writing a library crate, your crate must not enable any features of k8s-openapi directly. The choice of which feature to enable must be left to the application crate that uses your library. This ensures that all k8s-openapi-using dependencies in that application crate's dependency graph use the same set of k8s-openapi types and are interoperable.

If your library crate has tests or examples, you should also add a dev-dependency on k8s-openapi in addition to the direct dependency, and enable a version feature only for that dev-dependency.

# For library crates

[dependencies]
k8s-openapi = "..."

[dev-dependencies]
k8s-openapi = { version = "...", features = ["v1_14"] }

Your crate may need to emit different code depending on which feature of k8s-openapi gets selected eventually. For such things, use the k8s_* macros to emit different code depending on which feature eventually gets enabled. See the docs of the macros for more details.

For example:

Your crate requires the v1_11 or higher feature to be enabled because it uses types that were only added in Kubernetes 1.11. Your crate would fail to compile if a lower feature was enabled.

You can generate a custom compiler error with the compile_error! macro, and emit it from your crate root.

#[macro_use] extern crate k8s_openapi;

k8s_if_le_1_10! {
    compile_error!("This crate requires the v1_11 (or higher) feature to be enabled on the k8s-openapi crate.");
}

Your crate creates a custom resource definition. If the v1_9 or later feature is enabled, your crate also wants to set a validation spec on the CRD.

let custom_resource_definition_spec = apiextensions::CustomResourceDefinitionSpec {
    ... // Set all fields except `validation`, ie only those available in v1.8
};

k8s_if_ge_1_9! {
    // Set `validation`
    let custom_resource_definition_spec = apiextensions::CustomResourceDefinitionSpec {
        validation: Some(apiextensions::CustomResourceValidation { ... },
        ..custom_resource_definition_spec
    };
}

(These macros are required because cargo does not give any way for your crate to determine the features enabled on another crate, ie k8s-openapi.)

Examples

Resources

This example creates an instance of api::core::v1::PodSpec with no other properties set, and pretty-prints it.

use k8s_openapi::api::core::v1 as api;

fn main() {
    let pod_spec: api::PodSpec = Default::default();
    println!("{:#?}", pod_spec);
}

Client API

This example executes the [api::core::v1::Pod::list_namespaced_pod] API operation to list all pods inside a namespace. It demonstrates the common patterns implemented by all API operation functions in this crate:

The API function has required parameters and optional parameters. All optional parameters are taken as a single struct with optional fields.

Specifically for the [api::core::v1::Pod::list_namespaced_pod] operation, the namespace parameter is required and taken by the function itself, while other optional parameters like field_selector are fields of the ListOptional struct. An instance of this struct is taken as the last parameter of Pod::list_namespaced_pod. This struct impls Default so that you can just pass in Default::default() if you don't want to specify values for any of the optional parameters.

Some API operations have a single common type for optional parameters:

All create API take optional parameters using the CreateOptional struct.

All delete API take optional parameters using the DeleteOptional struct.

All list API take optional parameters using the ListOptional struct.

All patch API take optional parameters using the PatchOptional struct.

All replace API take optional parameters using the ReplaceOptional struct.

All watch API take optional parameters using the WatchOptional struct.

All delete-collection API take optional parameters using the DeleteOptional struct for delete options and the ListOptional struct for list options.

Other API functions have their own Optional structs with fields corresponding to the specific parameters for those functions, such as [api::core::v1::ReadNamespacedPodOptional] for [api::core::v1::Pod::read_namespaced_pod]

The function returns an http::Request value with the URL path, query string, and request body filled out according to the parameters given to the function. The function does not execute this request. You can execute this http::Request using any HTTP client library you want to use. It does not matter whether you use a synchronous client like reqwest, or an asynchronous client like hyper, or a mock client that returns bytes read from a test file.

For each API operation function, there is a corresponding response type. For Pod::list_namespaced_pod this is ListResponse<api::core::v1::Pod>. This is an enum with variants for each of the possible HTTP status codes that the operation can return, and contains the data that the API server would return corresponding to that status code. For example, the list-namespaced-pod operation returns a pod list with HTTP 200 OK, so one of the variants of that type is Ok(List<api::core::v1::Pod>)

The response types impl the Response trait, which contains a single Response::try_from_parts function. This function takes an http::StatusCode and a &u8 byte buffer, and tries to parse the byte buffer as the response type. For example, if you executed the request and received an HTTP 200 OK response with some bytes, you could call <ListResponse<Pod> as Response>::try_from_parts(status_code, buf) and expect to get Ok(ListResponse::<Pod>::Ok(pod_list)) from it.

Once again, this design ensures that the crate is not tied to a specific HTTP client library or interface. It does not matter how you execute the HTTP request, nor whether your library is synchronous or asynchronous, since every HTTP client library gives you a way to get the HTTP response status code and the bytes of the response body.

The API operation function also returns another value next to the http::Request. This value is a function that takes an http::StatusCode and returns a ResponseBody<ListResponse<Pod>>. As mentioned above, Response::try_from_parts requires you to maintain a byte buffer for the response body. ResponseBody is a helper that maintains such a buffer internally. It provides an append_slice() function to append slices to this internal buffer, and a parse() function to parse the buffer as the expected type (ListResponse<Pod> in this case).

It is not necessary to use the ResponseBody returned by the API operation function to parse the response. The ResponseBody::parse function is only a wrapper around the underlying Response::try_from_parts function, and handles growing and shrinking its inner buffer as necessary. It also helps ensure that the response body is parsed as the correct type for the operation, ListResponse<Pod> in this case, and not some other type. However, you can instead use your own byte buffer instead of the ResponseBody value and call ListResponse<Pod>::try_from_parts yourself.

The response types are enums with variants corresponding to HTTP status codes. For example, the ListResponse<Pod>::Ok variant corresponds to the HTTP 200 response of the list-namespaced-pod API.

Each response enum also has an Other variant, that is yielded when the response status code does not match any of the other variants. This variant has a Result<Option<serde_json::Value>, serde_json::Error> value.

If the response body is empty, this value will be Ok(None).

If the response body is not empty, this value will be an Ok(Some(value)) or Err(err) from attempting to parse that body as a serde_json::Value. If you expect the response body to be a specific JSON type such as apimachinery::pkg::apis::meta::v1::Status, you can use the serde_json::Value as a serde::Deserializer like let status = <Status as Deserialize>::deserialize(value)?;. On the other hand, if you expect the response body to not be a JSON value, then ignore the Err(err) and parse the raw bytes of the response into the appropriate type.

Also see the get_single_value and get_multiple_values functions in the k8s-openapi-tests directory in the repository for examples of how to use a synchronous client with this style of API.

// Re-export of the http crate since it's used in the public API
use k8s_openapi::http;

use k8s_openapi::api::core::v1 as api;

// Assume `execute` is some function that takes an `http::Request` and
// executes it synchronously or asynchronously to get a response. This is
// provided by your HTTP client library.
//
// Note that the `http::Request` values returned by API operation functions
// only have a URL path, query string and request body filled out. That is,
// they do *not* have a URL host. So the real `execute` implementation
// would first mutate the URL of the request to an absolute URL with
// the API server's authority, add authorization headers, etc before
// actually executing it.
fn execute(req: http::Request<Vec<u8>>) -> Response { unimplemented!(); }

fn main() -> Result<(), Box<dyn std::error::Error>> {
    // Create a `http::Request` to list all the pods in the
    // "kube-system" namespace.
    let (request, response_body) =
        api::Pod::list_namespaced_pod("kube-system", Default::default())?;

    // Execute the request and get a response.
    // If this is an asynchronous operation, you would await
    // or otherwise yield to the event loop here.
    let response = execute(request);

    // Got a status code from executing the request.
    let status_code: http::StatusCode = response.status_code();

    // Construct the `ResponseBody<ListResponse<Pod>>` using the
    // constructor returned by the API function.
    let mut response_body = response_body(status_code);

    // Buffer used for each read from the HTTP response.
    let mut buf = Box::new([0u8; 4096]);

    let pod_list = loop {
        // Read some bytes from the HTTP response into the buffer.
        // If this is an asynchronous operation, you would await or
        // yield to the event loop here.
        let read = response.read_into(&mut *buf)?;

        // `buf` now contains some data read from the response. Append it
        // to the `ResponseBody` and try to parse it into
        // the response type.
        response_body.append_slice(&buf[..read]);
        let response = response_body.parse();
        match response {
            // Successful response (HTTP 200 and parsed successfully)
            Ok(k8s_openapi::ListResponse::Ok(pod_list)) =>
                break pod_list,

            // Some unexpected response
            // (not HTTP 200, but still parsed successfully)
            Ok(other) => return Err(format!(
                "expected Ok but got {} {:?}",
                status_code, other).into()),

            // Need more response data.
            // Read more bytes from the response into the `ResponseBody`
            Err(k8s_openapi::ResponseError::NeedMoreData) => continue,

            // Some other error, like the response body being
            // malformed JSON or invalid UTF-8.
            Err(err) => return Err(format!(
                "error: {} {:?}",
                status_code, err).into()),
        }
    };

    for pod in pod_list.items {
        println!("{:#?}", pod);
    }

    Ok(())
}

Crate features

As mentioned above, enabling one of the v1_* features selects which version of the Kubernetes API server this crate should target.

If the api feature is disabled, the library will only contain the resource types like api::core::v1::Pod and not the associated operation functions like [api::core::v1::Pod::read_namespaced_pod]. The corresponding Response and Optional types will also not be accessible.

This feature is enabled by default, but can be disabled if your crate does not need the operation functions to save on compile time and resources.

Custom resource definitions

The k8s-openapi-derive crate provides a custom derive for generating clientsets for custom resources. See that crate's docs for more information.

Re-exports

pub use chrono;

pub use http;

pub use percent_encoding;

pub use serde_json;

pub use url;

Modules

api

apiextensions_apiserver

apimachinery

kube_aggregator

percent_encoding2
Extensions to the percent-encoding crate

Macros

k8s_match
A macro that emits a match expr with the given test expression and arms. The match arms can be annotated with the other conditional compilation macros in this crate so that they're only emitted if the predicate is true.

k8s_if_1_8
This macro evaluates to its contents if the v1_8 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_9
This macro evaluates to its contents if the v1_9 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_10
This macro evaluates to its contents if the v1_10 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_11
This macro evaluates to its contents if the v1_11 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_12
This macro evaluates to its contents if the v1_12 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_13
This macro evaluates to its contents if the v1_13 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_14
This macro evaluates to its contents if the v1_14 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_15
This macro evaluates to its contents if the v1_15 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_16
This macro evaluates to its contents if the v1_16 feature is enabled, otherwise it evaluates to nothing.

k8s_if_1_17
This macro evaluates to its contents if the v1_17 feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_8
This macro evaluates to its contents if the v1_8 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_9
This macro evaluates to its contents if the v1_9 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_10
This macro evaluates to its contents if the v1_10 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_11
This macro evaluates to its contents if the v1_11 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_12
This macro evaluates to its contents if the v1_12 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_13
This macro evaluates to its contents if the v1_13 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_14
This macro evaluates to its contents if the v1_14 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_15
This macro evaluates to its contents if the v1_15 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_16
This macro evaluates to its contents if the v1_16 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_ge_1_17
This macro evaluates to its contents if the v1_17 or higher feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_8
This macro evaluates to its contents if the v1_8 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_9
This macro evaluates to its contents if the v1_9 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_10
This macro evaluates to its contents if the v1_10 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_11
This macro evaluates to its contents if the v1_11 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_12
This macro evaluates to its contents if the v1_12 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_13
This macro evaluates to its contents if the v1_13 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_14
This macro evaluates to its contents if the v1_14 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_15
This macro evaluates to its contents if the v1_15 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_16
This macro evaluates to its contents if the v1_16 or lower feature is enabled, otherwise it evaluates to nothing.

k8s_if_le_1_17
This macro evaluates to its contents if the v1_17 or lower feature is enabled, otherwise it evaluates to nothing.

Structs

ByteString
A wrapper around a list of bytes.

CreateOptional
Common parameters for all create operations.

DeleteOptional
Common parameters for all delete and delete-collection operations.

List
List is a list of resources.

ListOptional
Common parameters for all list operations.

PatchOptional
Common parameters for all patch operations.

ReplaceOptional
Common parameters for all replace operations.

ResponseBody
This struct provides an easy way to parse a byte buffer into a Kubernetes API function's response.

WatchOptional
Common parameters for all watch operations.

Enums

CreateResponse
The common response type for all create API operations.

DeleteResponse
The common response type for all delete API operations and delete-collection API operations.

ListResponse
The common response type for all list API operations.

PatchResponse
The common response type for all patch API operations.

ReplaceResponse
The common response type for all replace API operations.

RequestError
The type of errors returned by the Kubernetes API functions that prepare the HTTP request.

ResponseError
The type of errors from parsing an HTTP response as one of the Kubernetes API functions' response types.

WatchResponse
The common response type for all watch API operations.

Traits

ListableResource
A trait applied to all Kubernetes resources that can be part of a corresponding list.

Metadata
A trait applied to all Kubernetes resources that have metadata.

Resource
A trait applied to all Kubernetes resources.

Response
A trait implemented by all response types corresponding to Kubernetes API functions.

Functions

api_version
Extracts the API version of the given resource value.

group
Extracts the group of the given resource value.

kind
Extracts the kind of the given resource value.

version
Extracts the version of the given resource value.