Struct k8s_openapi::api::admissionregistration::v1::ValidatingWebhookConfiguration

pub struct ValidatingWebhookConfiguration {
    pub metadata: ObjectMeta,
    pub webhooks: Option<Vec<ValidatingWebhook>>,
}

ValidatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and object without changing it.

Fields
metadata: ObjectMeta

Standard object metadata; More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata.

webhooks: Option<Vec<ValidatingWebhook>>

Webhooks is a list of webhooks and the affected resources and operations.

Implementations

impl ValidatingWebhookConfiguration

pub fn create(
body: &ValidatingWebhookConfiguration,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn delete_collection(
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn delete(
name: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ValidatingWebhookConfiguration

optional

Optional parameters. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn list(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind ValidatingWebhookConfiguration

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn patch(
name: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ValidatingWebhookConfiguration

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn read(
name: &str
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadValidatingWebhookConfigurationResponse>), RequestError>

read the specified ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<ReadValidatingWebhookConfigurationResponse> constructor, or ReadValidatingWebhookConfigurationResponse directly, to parse the HTTP response.

Arguments

name

name of the ValidatingWebhookConfiguration

impl ValidatingWebhookConfiguration

pub fn replace(
name: &str,
body: &ValidatingWebhookConfiguration,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified ValidatingWebhookConfiguration

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ValidatingWebhookConfiguration

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ValidatingWebhookConfiguration

pub fn watch(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind ValidatingWebhookConfiguration

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.