k8s_openapi::api::admissionregistration
Module v1beta1

Structs

AuditAnnotation

AuditAnnotation describes how to produce an audit annotation for an API request.

ExpressionWarning

ExpressionWarning is a warning information that targets a specific expression.

MatchCondition

MatchCondition represents a condition which must be fulfilled for a request to be sent to a webhook.

MatchResources

MatchResources decides whether to run the admission control policy on an object based on whether it meets the match criteria. The exclude rules take precedence over include rules (if a resource matches both, it is excluded)

NamedRuleWithOperations

NamedRuleWithOperations is a tuple of Operations and Resources with ResourceNames.

ParamKind

ParamKind is a tuple of Group Kind and Version.

ParamRef

ParamRef describes how to locate the params to be used as input to expressions of rules applied by a policy binding.

TypeChecking

TypeChecking contains results of type checking the expressions in the ValidatingAdmissionPolicy

ValidatingAdmissionPolicy

ValidatingAdmissionPolicy describes the definition of an admission validation policy that accepts or rejects an object without changing it.

ValidatingAdmissionPolicyBinding

ValidatingAdmissionPolicyBinding binds the ValidatingAdmissionPolicy with paramerized resources. ValidatingAdmissionPolicyBinding and parameter CRDs together define how cluster administrators configure policies for clusters.

ValidatingAdmissionPolicyBindingSpec

ValidatingAdmissionPolicyBindingSpec is the specification of the ValidatingAdmissionPolicyBinding.

ValidatingAdmissionPolicySpec

ValidatingAdmissionPolicySpec is the specification of the desired behavior of the AdmissionPolicy.

ValidatingAdmissionPolicyStatus

ValidatingAdmissionPolicyStatus represents the status of an admission validation policy.

Validation

Validation specifies the CEL expression which is used to apply the validation.

Variable

Variable is the definition of a variable that is used for composition. A variable is defined as a named expression.