Struct k8s_openapi::api::apps::v1::ControllerRevision

pub struct ControllerRevision {
    pub data: Option<RawExtension>,
    pub metadata: ObjectMeta,
    pub revision: i64,
}

ControllerRevision implements an immutable snapshot of state data. Clients are responsible for serializing and deserializing the objects that contain their internal state. Once a ControllerRevision has been successfully created, it can not be updated. The API Server will fail validation of all requests that attempt to mutate the Data field. ControllerRevisions may, however, be deleted. Note that, due to its use by both the DaemonSet and StatefulSet controllers for update and rollback, this object is beta. However, it may be subject to name and representation changes in future releases, and clients should not depend on its stability. It is primarily for internal use by controllers.

Fields
data: Option<RawExtension>

Data is the serialized representation of the state.

metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

revision: i64

Revision indicates the revision of the state represented by Data.

Implementations

impl ControllerRevision

pub fn create(
namespace: &str,
body: &ControllerRevision,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a ControllerRevision

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn delete_collection(
namespace: &str,
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of ControllerRevision

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl ControllerRevision

pub fn delete(
name: &str,
namespace: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a ControllerRevision

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ControllerRevision

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn list_for_all_namespaces(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind ControllerRevision

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn list(
namespace: &str,
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind ControllerRevision

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn patch(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified ControllerRevision

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ControllerRevision

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn read(
name: &str,
namespace: &str
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadControllerRevisionResponse>), RequestError>

read the specified ControllerRevision

Use the returned crate::ResponseBody<ReadControllerRevisionResponse> constructor, or ReadControllerRevisionResponse directly, to parse the HTTP response.

Arguments

name

name of the ControllerRevision

namespace

object name and auth scope, such as for teams and projects

impl ControllerRevision

pub fn replace(
name: &str,
namespace: &str,
body: &ControllerRevision,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified ControllerRevision

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the ControllerRevision

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn watch_for_all_namespaces(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind ControllerRevision

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl ControllerRevision

pub fn watch(
namespace: &str,
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind ControllerRevision

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.