Struct k8s_openapi::api::apps::v1::DaemonSet

pub struct DaemonSet {
    pub metadata: ObjectMeta,
    pub spec: Option<DaemonSetSpec>,
    pub status: Option<DaemonSetStatus>,
}

DaemonSet represents the configuration of a daemon set.

Fields
metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<DaemonSetSpec>

The desired behavior of this daemon set. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

status: Option<DaemonSetStatus>

The current status of this daemon set. This data may be out of date by some window of time. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

Implementations

impl DaemonSet

pub fn create(
namespace: &str,
body: &DaemonSet,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a DaemonSet

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn delete_collection(
namespace: &str,
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of DaemonSet

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl DaemonSet

pub fn delete(
name: &str,
namespace: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a DaemonSet

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn list_for_all_namespaces(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind DaemonSet

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn list(
namespace: &str,
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind DaemonSet

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn patch(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified DaemonSet

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn patch_status(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified DaemonSet

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn read(
name: &str,
namespace: &str
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadDaemonSetResponse>), RequestError>

read the specified DaemonSet

Use the returned crate::ResponseBody<ReadDaemonSetResponse> constructor, or ReadDaemonSetResponse directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

impl DaemonSet

pub fn read_status(
name: &str,
namespace: &str
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadDaemonSetStatusResponse>), RequestError>

read status of the specified DaemonSet

Use the returned crate::ResponseBody<ReadDaemonSetStatusResponse> constructor, or ReadDaemonSetStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

impl DaemonSet

pub fn replace(
name: &str,
namespace: &str,
body: &DaemonSet,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified DaemonSet

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn replace_status(
name: &str,
namespace: &str,
body: &DaemonSet,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified DaemonSet

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the DaemonSet

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn watch_for_all_namespaces(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind DaemonSet

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl DaemonSet

pub fn watch(
namespace: &str,
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind DaemonSet

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.