Module k8s_openapi::api::auditregistration::v1alpha1

Structs

AuditSink
AuditSink represents a cluster level audit sink

AuditSinkSpec
AuditSinkSpec holds the spec for the audit sink

Policy
Policy defines the configuration of how audit events are logged

ReadAuditSinkOptional
Optional parameters of AuditSink::read_audit_sink

ServiceReference
ServiceReference holds a reference to Service.legacy.k8s.io

Webhook
Webhook holds the configuration of the webhook

WebhookClientConfig
WebhookClientConfig contains the information to make a connection with the webhook

WebhookThrottleConfig
WebhookThrottleConfig holds the configuration for throttling events

Enums

ReadAuditSinkResponse
Use <ReadAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::read_audit_sink