k8s_openapi::api::authentication
Module v1

Structs

BoundObjectReference

BoundObjectReference is a reference to an object that a token is bound to.

SelfSubjectReview

SelfSubjectReview contains the user information that the kube-apiserver has about the user making this request. When using impersonation, users will receive the user info of the user being impersonated. If impersonation or request header authentication is used, any extra keys will have their case ignored and returned as lowercase.

SelfSubjectReviewStatus

SelfSubjectReviewStatus is filled by the kube-apiserver and sent back to a user.

TokenRequest

TokenRequest requests a token for a given service account.

TokenRequestSpec

TokenRequestSpec contains client provided parameters of a token request.

TokenRequestStatus

TokenRequestStatus is the result of a token request.

TokenReview

TokenReview attempts to authenticate a token to a known user. Note: TokenReview requests may be cached by the webhook token authenticator plugin in the kube-apiserver.

TokenReviewSpec

TokenReviewSpec is a description of the token authentication request.

TokenReviewStatus

TokenReviewStatus is the result of the token authentication request.

UserInfo

UserInfo holds the information about the user needed to implement the user.Info interface.