Module k8s_openapi::api::autoscaling::v1

Structs

CrossVersionObjectReference

CrossVersionObjectReference contains enough information to let you identify the referred resource.

HorizontalPodAutoscaler

configuration of a horizontal pod autoscaler.

HorizontalPodAutoscalerSpec

specification of a horizontal pod autoscaler.

HorizontalPodAutoscalerStatus

current status of a horizontal pod autoscaler

Scale

Scale represents a scaling request for a resource.

ScaleSpec

ScaleSpec describes the attributes of a scale subresource.

ScaleStatus

ScaleStatus represents the current status of a scale subresource.

Enums

ReadDeploymentScaleResponse

Use <ReadDeploymentScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_deployment

ReadHorizontalPodAutoscalerResponse

Use <ReadHorizontalPodAutoscalerResponse as Response>::try_from_parts to parse the HTTP response body of HorizontalPodAutoscaler::read

ReadHorizontalPodAutoscalerStatusResponse

Use <ReadHorizontalPodAutoscalerStatusResponse as Response>::try_from_parts to parse the HTTP response body of HorizontalPodAutoscaler::read_status

ReadReplicaSetScaleResponse

Use <ReadReplicaSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_replica_set

ReadReplicationControllerScaleResponse

Use <ReadReplicationControllerScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_replication_controller

ReadStatefulSetScaleResponse

Use <ReadStatefulSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_stateful_set