Struct k8s_openapi::api::autoscaling::v1::HorizontalPodAutoscaler

pub struct HorizontalPodAutoscaler {
    pub metadata: ObjectMeta,
    pub spec: Option<HorizontalPodAutoscalerSpec>,
    pub status: Option<HorizontalPodAutoscalerStatus>,
}

configuration of a horizontal pod autoscaler.

Fields
metadata: ObjectMeta

Standard object metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<HorizontalPodAutoscalerSpec>

behaviour of autoscaler. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.

status: Option<HorizontalPodAutoscalerStatus>

current information about the autoscaler.

Implementations

impl HorizontalPodAutoscaler

pub fn create_namespaced_horizontal_pod_autoscaler(
namespace: &str,
body: &HorizontalPodAutoscaler,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn delete_collection_namespaced_horizontal_pod_autoscaler(
namespace: &str,
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn delete_namespaced_horizontal_pod_autoscaler(
name: &str,
namespace: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn list_horizontal_pod_autoscaler_for_all_namespaces(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind HorizontalPodAutoscaler

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn list_namespaced_horizontal_pod_autoscaler(
namespace: &str,
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind HorizontalPodAutoscaler

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn patch_namespaced_horizontal_pod_autoscaler(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn patch_namespaced_horizontal_pod_autoscaler_status(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn read_namespaced_horizontal_pod_autoscaler(
name: &str,
namespace: &str,
optional: ReadNamespacedHorizontalPodAutoscalerOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedHorizontalPodAutoscalerResponse>), RequestError>

read the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<ReadNamespacedHorizontalPodAutoscalerResponse> constructor, or ReadNamespacedHorizontalPodAutoscalerResponse directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn read_namespaced_horizontal_pod_autoscaler_status(
name: &str,
namespace: &str,
optional: ReadNamespacedHorizontalPodAutoscalerStatusOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedHorizontalPodAutoscalerStatusResponse>), RequestError>

read status of the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<ReadNamespacedHorizontalPodAutoscalerStatusResponse> constructor, or ReadNamespacedHorizontalPodAutoscalerStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn replace_namespaced_horizontal_pod_autoscaler(
name: &str,
namespace: &str,
body: &HorizontalPodAutoscaler,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn replace_namespaced_horizontal_pod_autoscaler_status(
name: &str,
namespace: &str,
body: &HorizontalPodAutoscaler,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified HorizontalPodAutoscaler

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the HorizontalPodAutoscaler

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn watch_horizontal_pod_autoscaler_for_all_namespaces(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind HorizontalPodAutoscaler

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl HorizontalPodAutoscaler

pub fn watch_namespaced_horizontal_pod_autoscaler(
namespace: &str,
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind HorizontalPodAutoscaler

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.