Struct k8s_openapi::api::autoscaling::v1::Scale

pub struct Scale {
    pub metadata: Option<ObjectMeta>,
    pub spec: Option<ScaleSpec>,
    pub status: Option<ScaleStatus>,
}

Scale represents a scaling request for a resource.

Fields
metadata: Option<ObjectMeta>

Standard object metadata; More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata.

spec: Option<ScaleSpec>

defines the behavior of the scale. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.

status: Option<ScaleStatus>

current status of the scale. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status. Read-only.

Methods

impl Scale

pub fn patch_namespaced_deployment_scale(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedDeploymentScaleResponse>), RequestError>

partially update scale of the specified Deployment

Use the returned crate::ResponseBody<PatchNamespacedDeploymentScaleResponse> constructor, or PatchNamespacedDeploymentScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_namespaced_replica_set_scale(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedReplicaSetScaleResponse>), RequestError>

partially update scale of the specified ReplicaSet

Use the returned crate::ResponseBody<PatchNamespacedReplicaSetScaleResponse> constructor, or PatchNamespacedReplicaSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_namespaced_stateful_set_scale(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedStatefulSetScaleResponse>), RequestError>

partially update scale of the specified StatefulSet

Use the returned crate::ResponseBody<PatchNamespacedStatefulSetScaleResponse> constructor, or PatchNamespacedStatefulSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn patch_namespaced_replication_controller_scale(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedReplicationControllerScaleResponse>), RequestError>

partially update scale of the specified ReplicationController

Use the returned crate::ResponseBody<PatchNamespacedReplicationControllerScaleResponse> constructor, or PatchNamespacedReplicationControllerScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn read_namespaced_deployment_scale(
name: &str,
namespace: &str,
optional: ReadNamespacedDeploymentScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedDeploymentScaleResponse>), RequestError>

read scale of the specified Deployment

Use the returned crate::ResponseBody<ReadNamespacedDeploymentScaleResponse> constructor, or ReadNamespacedDeploymentScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn read_namespaced_replica_set_scale(
name: &str,
namespace: &str,
optional: ReadNamespacedReplicaSetScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedReplicaSetScaleResponse>), RequestError>

read scale of the specified ReplicaSet

Use the returned crate::ResponseBody<ReadNamespacedReplicaSetScaleResponse> constructor, or ReadNamespacedReplicaSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn read_namespaced_stateful_set_scale(
name: &str,
namespace: &str,
optional: ReadNamespacedStatefulSetScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedStatefulSetScaleResponse>), RequestError>

read scale of the specified StatefulSet

Use the returned crate::ResponseBody<ReadNamespacedStatefulSetScaleResponse> constructor, or ReadNamespacedStatefulSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn read_namespaced_replication_controller_scale(
name: &str,
namespace: &str,
optional: ReadNamespacedReplicationControllerScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedReplicationControllerScaleResponse>), RequestError>

read scale of the specified ReplicationController

Use the returned crate::ResponseBody<ReadNamespacedReplicationControllerScaleResponse> constructor, or ReadNamespacedReplicationControllerScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_namespaced_deployment_scale(
name: &str,
namespace: &str,
body: &Scale,
optional: ReplaceNamespacedDeploymentScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedDeploymentScaleResponse>), RequestError>

replace scale of the specified Deployment

Use the returned crate::ResponseBody<ReplaceNamespacedDeploymentScaleResponse> constructor, or ReplaceNamespacedDeploymentScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_namespaced_replica_set_scale(
name: &str,
namespace: &str,
body: &Scale,
optional: ReplaceNamespacedReplicaSetScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedReplicaSetScaleResponse>), RequestError>

replace scale of the specified ReplicaSet

Use the returned crate::ResponseBody<ReplaceNamespacedReplicaSetScaleResponse> constructor, or ReplaceNamespacedReplicaSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_namespaced_stateful_set_scale(
name: &str,
namespace: &str,
body: &Scale,
optional: ReplaceNamespacedStatefulSetScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedStatefulSetScaleResponse>), RequestError>

replace scale of the specified StatefulSet

Use the returned crate::ResponseBody<ReplaceNamespacedStatefulSetScaleResponse> constructor, or ReplaceNamespacedStatefulSetScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Scale

pub fn replace_namespaced_replication_controller_scale(
name: &str,
namespace: &str,
body: &Scale,
optional: ReplaceNamespacedReplicationControllerScaleOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedReplicationControllerScaleResponse>), RequestError>

replace scale of the specified ReplicationController

Use the returned crate::ResponseBody<ReplaceNamespacedReplicationControllerScaleResponse> constructor, or ReplaceNamespacedReplicationControllerScaleResponse directly, to parse the HTTP response.

Arguments

name

name of the Scale

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.