Struct k8s_openapi::api::batch::v1::JobStatus

pub struct JobStatus {
    pub active: Option<i32>,
    pub completed_indexes: Option<String>,
    pub completion_time: Option<Time>,
    pub conditions: Option<Vec<JobCondition>>,
    pub failed: Option<i32>,
    pub failed_indexes: Option<String>,
    pub ready: Option<i32>,
    pub start_time: Option<Time>,
    pub succeeded: Option<i32>,
    pub terminating: Option<i32>,
    pub uncounted_terminated_pods: Option<UncountedTerminatedPods>,
}

JobStatus represents the current state of a Job.

Fields
active: Option<i32>

The number of pending and running pods.

completed_indexes: Option<String>

completedIndexes holds the completed indexes when .spec.completionMode = โIndexedโ in a text format. The indexes are represented as decimal integers separated by commas. The numbers are listed in increasing order. Three or more consecutive numbers are compressed and represented by the first and last element of the series, separated by a hyphen. For example, if the completed indexes are 1, 3, 4, 5 and 7, they are represented as โ1,3-5,7โ.

completion_time: Option<Time>

Represents time when the job was completed. It is not guaranteed to be set in happens-before order across separate operations. It is represented in RFC3339 form and is in UTC. The completion time is only set when the job finishes successfully.

conditions: Option<Vec<JobCondition>>

The latest available observations of an objectโs current state. When a Job fails, one of the conditions will have type โFailedโ and status true. When a Job is suspended, one of the conditions will have type โSuspendedโ and status true; when the Job is resumed, the status of this condition will become false. When a Job is completed, one of the conditions will have type โCompleteโ and status true. More info: https://kubernetes.io/docs/concepts/workloads/controllers/jobs-run-to-completion/

failed: Option<i32>

The number of pods which reached phase Failed.

failed_indexes: Option<String>

FailedIndexes holds the failed indexes when backoffLimitPerIndex=true. The indexes are represented in the text format analogous as for the completedIndexes field, ie. they are kept as decimal integers separated by commas. The numbers are listed in increasing order. Three or more consecutive numbers are compressed and represented by the first and last element of the series, separated by a hyphen. For example, if the failed indexes are 1, 3, 4, 5 and 7, they are represented as โ1,3-5,7โ. This field is alpha-level. It can be used when the JobBackoffLimitPerIndex feature gate is enabled (disabled by default).

ready: Option<i32>

The number of pods which have a Ready condition.

This field is beta-level. The job controller populates the field when the feature gate JobReadyPods is enabled (enabled by default).

start_time: Option<Time>

Represents time when the job controller started processing a job. When a Job is created in the suspended state, this field is not set until the first time it is resumed. This field is reset every time a Job is resumed from suspension. It is represented in RFC3339 form and is in UTC.

succeeded: Option<i32>

The number of pods which reached phase Succeeded.

terminating: Option<i32>

The number of pods which are terminating (in phase Pending or Running and have a deletionTimestamp).

This field is alpha-level. The job controller populates the field when the feature gate JobPodReplacementPolicy is enabled (disabled by default).

uncounted_terminated_pods: Option<UncountedTerminatedPods>

uncountedTerminatedPods holds the UIDs of Pods that have terminated but the job controller hasnโt yet accounted for in the status counters.

The job controller creates pods with a finalizer. When a pod terminates (succeeded or failed), the controller does three steps to account for it in the job status:

Add the pod UID to the arrays in this field. 2. Remove the pod finalizer. 3. Remove the pod UID from the arrays while increasing the corresponding counter.

Old jobs might not be tracked using this field, in which case the field remains null.