k8s_openapi::api::certificates

Module v1beta1

Structs

ClusterTrustBundle

ClusterTrustBundle is a cluster-scoped container for X.509 trust anchors (root certificates).

ClusterTrustBundleSpec

ClusterTrustBundleSpec contains the signer and trust anchors.

PodCertificateRequest

PodCertificateRequest encodes a pod requesting a certificate from a given signer.

PodCertificateRequestSpec

PodCertificateRequestSpec describes the certificate request. All fields are immutable after creation.

PodCertificateRequestStatus

PodCertificateRequestStatus describes the status of the request, and holds the certificate data if the request is issued.