Struct k8s_openapi::api::core::v1::ContainerStatus

pub struct ContainerStatus {
    pub allocated_resources: Option<BTreeMap<String, Quantity>>,
    pub container_id: Option<String>,
    pub image: String,
    pub image_id: String,
    pub last_state: Option<ContainerState>,
    pub name: String,
    pub ready: bool,
    pub resources: Option<ResourceRequirements>,
    pub restart_count: i32,
    pub started: Option<bool>,
    pub state: Option<ContainerState>,
}

ContainerStatus contains details for the current status of this container.

Fields
allocated_resources: Option<BTreeMap<String, Quantity>>

AllocatedResources represents the compute resources allocated for this container by the node. Kubelet sets this value to Container.Resources.Requests upon successful pod admission and after successfully admitting desired pod resize.

container_id: Option<String>

ContainerID is the ID of the container in the format โ<type>://<container_id>โ. Where type is a container runtime identifier, returned from Version call of CRI API (for example โcontainerdโ).

image: String

Image is the name of container image that the container is running. The container image may not match the image used in the PodSpec, as it may have been resolved by the runtime. More info: https://kubernetes.io/docs/concepts/containers/images.

image_id: String

ImageID is the image ID of the containerโs image. The image ID may not match the image ID of the image used in the PodSpec, as it may have been resolved by the runtime.

last_state: Option<ContainerState>

LastTerminationState holds the last termination state of the container to help debug container crashes and restarts. This field is not populated if the container is still running and RestartCount is 0.

name: String

Name is a DNS_LABEL representing the unique name of the container. Each container in a pod must have a unique name across all container types. Cannot be updated.

ready: bool

Ready specifies whether the container is currently passing its readiness check. The value will change as readiness probes keep executing. If no readiness probes are specified, this field defaults to true once the container is fully started (see Started field).

The value is typically used to determine whether a container is ready to accept traffic.

resources: Option<ResourceRequirements>

Resources represents the compute resource requests and limits that have been successfully enacted on the running container after it has been started or has been successfully resized.

restart_count: i32

RestartCount holds the number of times the container has been restarted. Kubelet makes an effort to always increment the value, but there are cases when the state may be lost due to node restarts and then the value may be reset to 0. The value is never negative.

started: Option<bool>

Started indicates whether the container has finished its postStart lifecycle hook and passed its startup probe. Initialized as false, becomes true after startupProbe is considered successful. Resets to false when the container is restarted, or if kubelet loses state temporarily. In both cases, startup probes will run again. Is always true when no startupProbe is defined and container is running and has passed the postStart lifecycle hook. The null value must be treated the same as false.

state: Option<ContainerState>

State holds details about the containerโs current condition.