Struct k8s_openapi::api::core::v1::Event

pub struct Event {
    pub action: Option<String>,
    pub count: Option<i32>,
    pub event_time: Option<MicroTime>,
    pub first_timestamp: Option<Time>,
    pub involved_object: ObjectReference,
    pub last_timestamp: Option<Time>,
    pub message: Option<String>,
    pub metadata: ObjectMeta,
    pub reason: Option<String>,
    pub related: Option<ObjectReference>,
    pub reporting_component: Option<String>,
    pub reporting_instance: Option<String>,
    pub series: Option<EventSeries>,
    pub source: Option<EventSource>,
    pub type_: Option<String>,
}

Event is a report of an event somewhere in the cluster.

Fields
action: Option<String>

What action was taken/failed regarding to the Regarding object.

count: Option<i32>

The number of times this event has occurred.

event_time: Option<MicroTime>

Time when this Event was first observed.

first_timestamp: Option<Time>

The time at which the event was first recorded. (Time of server receipt is in TypeMeta.)

involved_object: ObjectReference

The object that this event is about.

last_timestamp: Option<Time>

The time at which the most recent occurrence of this event was recorded.

message: Option<String>

A human-readable description of the status of this operation.

metadata: ObjectMeta

Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

reason: Option<String>

This should be a short, machine understandable string that gives the reason for the transition into the object's current status.

related: Option<ObjectReference>

Optional secondary object for more complex actions.

reporting_component: Option<String>

Name of the controller that emitted this Event, e.g. kubernetes.io/kubelet.

reporting_instance: Option<String>

ID of the controller instance, e.g. kubelet-xyzf.

series: Option<EventSeries>

Data about the Event series this event represents or nil if it's a singleton Event.

source: Option<EventSource>

The component reporting this event. Should be a short machine understandable string.

type_: Option<String>

Type of this event (Normal, Warning), new types could be added in the future

Methods

impl Event

pub fn create_namespaced_event(
namespace: &str,
body: &Event,
optional: CreateOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create an Event

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn delete_collection_namespaced_event(
namespace: &str,
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of Event

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Event

pub fn delete_namespaced_event(
name: &str,
namespace: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete an Event

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn list_event_for_all_namespaces(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Event

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or [crate::ListResponse]`` directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn list_namespaced_event(
namespace: &str,
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Event

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or [crate::ListResponse]`` directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn patch_namespaced_event(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified Event

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn read_namespaced_event(
name: &str,
namespace: &str,
optional: ReadNamespacedEventOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedEventResponse>), RequestError>

read the specified Event

Use the returned crate::ResponseBody<ReadNamespacedEventResponse> constructor, or ReadNamespacedEventResponse directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn replace_namespaced_event(
name: &str,
namespace: &str,
body: &Event,
optional: ReplaceOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified Event

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn watch_event_for_all_namespaces(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Event

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn watch_namespaced_event(
namespace: &str,
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Event

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.