Struct k8s_openapi::api::core::v1::PersistentVolume

pub struct PersistentVolume {
    pub metadata: ObjectMeta,
    pub spec: Option<PersistentVolumeSpec>,
    pub status: Option<PersistentVolumeStatus>,
}

PersistentVolume (PV) is a storage resource provisioned by an administrator. It is analogous to a node. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes

Fields
metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<PersistentVolumeSpec>

spec defines a specification of a persistent volume owned by the cluster. Provisioned by an administrator. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistent-volumes

status: Option<PersistentVolumeStatus>

status represents the current information/status for the persistent volume. Populated by the system. Read-only. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistent-volumes

Implementations

impl PersistentVolume

pub fn create( body: &PersistentVolume, optional: CreateOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a PersistentVolume

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn delete_collection( delete_optional: DeleteOptional<'_>, list_optional: ListOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of PersistentVolume

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl PersistentVolume

pub fn delete( name: &str, optional: DeleteOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a PersistentVolume

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn list( optional: ListOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind PersistentVolume

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn patch( name: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified PersistentVolume

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn patch_status( name: &str, body: &Patch, optional: PatchOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified PersistentVolume

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn read( name: &str ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadPersistentVolumeResponse>), RequestError>

read the specified PersistentVolume

Use the returned crate::ResponseBody<ReadPersistentVolumeResponse> constructor, or ReadPersistentVolumeResponse directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

impl PersistentVolume

pub fn read_status( name: &str ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadPersistentVolumeStatusResponse>), RequestError>

read status of the specified PersistentVolume

Use the returned crate::ResponseBody<ReadPersistentVolumeStatusResponse> constructor, or ReadPersistentVolumeStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

impl PersistentVolume

pub fn replace( name: &str, body: &PersistentVolume, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified PersistentVolume

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn replace_status( name: &str, body: &PersistentVolume, optional: ReplaceOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified PersistentVolume

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PersistentVolume

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PersistentVolume

pub fn watch( optional: WatchOptional<'_> ) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind PersistentVolume

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.