Struct k8s_openapi::api::core::v1::Pod

pub struct Pod {
    pub metadata: ObjectMeta,
    pub spec: Option<PodSpec>,
    pub status: Option<PodStatus>,
}

Pod is a collection of containers that can run on a host. This resource is created by clients and scheduled onto hosts.

Fields
metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<PodSpec>

Specification of the desired behavior of the pod. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

status: Option<PodStatus>

Most recently observed status of the pod. This data may not be up to date. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

Implementations

impl Pod

pub fn connect_delete_namespaced_pod_proxy(
name: &str,
namespace: &str,
optional: ConnectDeleteNamespacedPodProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_delete_namespaced_pod_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectDeleteNamespacedPodProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_get_namespaced_pod_attach(
name: &str,
namespace: &str,
optional: ConnectGetNamespacedPodAttachOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to attach of Pod

Arguments

name

name of the PodAttachOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_get_namespaced_pod_exec(
name: &str,
namespace: &str,
optional: ConnectGetNamespacedPodExecOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to exec of Pod

Arguments

name

name of the PodExecOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_get_namespaced_pod_portforward(
name: &str,
namespace: &str,
optional: ConnectGetNamespacedPodPortforwardOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to portforward of Pod

Arguments

name

name of the PodPortForwardOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_get_namespaced_pod_proxy(
name: &str,
namespace: &str,
optional: ConnectGetNamespacedPodProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_get_namespaced_pod_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectGetNamespacedPodProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_patch_namespaced_pod_proxy(
name: &str,
namespace: &str,
optional: ConnectPatchNamespacedPodProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_patch_namespaced_pod_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPatchNamespacedPodProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_post_namespaced_pod_attach(
name: &str,
namespace: &str,
optional: ConnectPostNamespacedPodAttachOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to attach of Pod

Arguments

name

name of the PodAttachOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_post_namespaced_pod_exec(
name: &str,
namespace: &str,
optional: ConnectPostNamespacedPodExecOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to exec of Pod

Arguments

name

name of the PodExecOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_post_namespaced_pod_portforward(
name: &str,
namespace: &str,
optional: ConnectPostNamespacedPodPortforwardOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to portforward of Pod

Arguments

name

name of the PodPortForwardOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_post_namespaced_pod_proxy(
name: &str,
namespace: &str,
optional: ConnectPostNamespacedPodProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_post_namespaced_pod_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPostNamespacedPodProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_put_namespaced_pod_proxy(
name: &str,
namespace: &str,
optional: ConnectPutNamespacedPodProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn connect_put_namespaced_pod_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPutNamespacedPodProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Pod

Arguments

name

name of the PodProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn create_namespaced_pod(
namespace: &str,
body: &Pod,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a Pod

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn delete_collection_namespaced_pod(
namespace: &str,
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of Pod

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Pod

pub fn delete_namespaced_pod(
name: &str,
namespace: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a Pod

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn list_namespaced_pod(
namespace: &str,
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Pod

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn list_pod_for_all_namespaces(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Pod

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn patch_namespaced_pod(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified Pod

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn patch_namespaced_pod_status(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified Pod

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn read_namespaced_pod(
name: &str,
namespace: &str,
optional: ReadNamespacedPodOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedPodResponse>), RequestError>

read the specified Pod

Use the returned crate::ResponseBody<ReadNamespacedPodResponse> constructor, or ReadNamespacedPodResponse directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn read_namespaced_pod_log(
name: &str,
namespace: &str,
optional: ReadNamespacedPodLogOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedPodLogResponse>), RequestError>

read log of the specified Pod

Use the returned crate::ResponseBody<ReadNamespacedPodLogResponse> constructor, or ReadNamespacedPodLogResponse directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn read_namespaced_pod_status(
name: &str,
namespace: &str,
optional: ReadNamespacedPodStatusOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedPodStatusResponse>), RequestError>

read status of the specified Pod

Use the returned crate::ResponseBody<ReadNamespacedPodStatusResponse> constructor, or ReadNamespacedPodStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn replace_namespaced_pod(
name: &str,
namespace: &str,
body: &Pod,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified Pod

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn replace_namespaced_pod_status(
name: &str,
namespace: &str,
body: &Pod,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified Pod

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Pod

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn watch_namespaced_pod(
namespace: &str,
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Pod

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Pod

pub fn watch_pod_for_all_namespaces(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Pod

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.