Struct k8s_openapi::api::core::v1::PodSecurityContext

pub struct PodSecurityContext {
    pub fs_group: Option<i64>,
    pub fs_group_change_policy: Option<String>,
    pub run_as_group: Option<i64>,
    pub run_as_non_root: Option<bool>,
    pub run_as_user: Option<i64>,
    pub se_linux_options: Option<SELinuxOptions>,
    pub seccomp_profile: Option<SeccompProfile>,
    pub supplemental_groups: Option<Vec<i64>>,
    pub sysctls: Option<Vec<Sysctl>>,
    pub windows_options: Option<WindowsSecurityContextOptions>,
}

PodSecurityContext holds pod-level security attributes and common container settings. Some fields are also present in container.securityContext. Field values of container.securityContext take precedence over field values of PodSecurityContext.

Fields
fs_group: Option<i64>

A special supplemental group that applies to all containers in a pod. Some volume types allow the Kubelet to change the ownership of that volume to be owned by the pod:

The owning GID will be the FSGroup 2. The setgid bit is set (new files created in the volume will be owned by FSGroup) 3. The permission bits are ORโd with rw-rwโโ

If unset, the Kubelet will not modify the ownership and permissions of any volume. Note that this field cannot be set when spec.os.name is windows.

fs_group_change_policy: Option<String>

fsGroupChangePolicy defines behavior of changing ownership and permission of the volume before being exposed inside Pod. This field will only apply to volume types which support fsGroup based ownership(and permissions). It will have no effect on ephemeral volume types such as: secret, configmaps and emptydir. Valid values are โOnRootMismatchโ and โAlwaysโ. If not specified, โAlwaysโ is used. Note that this field cannot be set when spec.os.name is windows.

run_as_group: Option<i64>

The GID to run the entrypoint of the container process. Uses runtime default if unset. May also be set in SecurityContext. If set in both SecurityContext and PodSecurityContext, the value specified in SecurityContext takes precedence for that container. Note that this field cannot be set when spec.os.name is windows.

run_as_non_root: Option<bool>

Indicates that the container must run as a non-root user. If true, the Kubelet will validate the image at runtime to ensure that it does not run as UID 0 (root) and fail to start the container if it does. If unset or false, no such validation will be performed. May also be set in SecurityContext. If set in both SecurityContext and PodSecurityContext, the value specified in SecurityContext takes precedence.

run_as_user: Option<i64>

The UID to run the entrypoint of the container process. Defaults to user specified in image metadata if unspecified. May also be set in SecurityContext. If set in both SecurityContext and PodSecurityContext, the value specified in SecurityContext takes precedence for that container. Note that this field cannot be set when spec.os.name is windows.

se_linux_options: Option<SELinuxOptions>

The SELinux context to be applied to all containers. If unspecified, the container runtime will allocate a random SELinux context for each container. May also be set in SecurityContext. If set in both SecurityContext and PodSecurityContext, the value specified in SecurityContext takes precedence for that container. Note that this field cannot be set when spec.os.name is windows.

seccomp_profile: Option<SeccompProfile>

The seccomp options to use by the containers in this pod. Note that this field cannot be set when spec.os.name is windows.

supplemental_groups: Option<Vec<i64>>

A list of groups applied to the first process run in each container, in addition to the containerโs primary GID, the fsGroup (if specified), and group memberships defined in the container image for the uid of the container process. If unspecified, no additional groups are added to any container. Note that group memberships defined in the container image for the uid of the container process are still effective, even if they are not included in this list. Note that this field cannot be set when spec.os.name is windows.

sysctls: Option<Vec<Sysctl>>

Sysctls hold a list of namespaced sysctls used for the pod. Pods with unsupported sysctls (by the container runtime) might fail to launch. Note that this field cannot be set when spec.os.name is windows.

windows_options: Option<WindowsSecurityContextOptions>

The Windows specific settings applied to all containers. If unspecified, the options within a containerโs SecurityContext will be used. If set in both SecurityContext and PodSecurityContext, the value specified in SecurityContext takes precedence. Note that this field cannot be set when spec.os.name is linux.