Struct k8s_openapi::api::core::v1::PodSpec

pub struct PodSpec {

    pub active_deadline_seconds: Option<i64>,
    pub affinity: Option<Affinity>,
    pub automount_service_account_token: Option<bool>,
    pub containers: Vec<Container>,
    pub dns_config: Option<PodDNSConfig>,
    pub dns_policy: Option<String>,
    pub enable_service_links: Option<bool>,
    pub ephemeral_containers: Option<Vec<EphemeralContainer>>,
    pub host_aliases: Option<Vec<HostAlias>>,
    pub host_ipc: Option<bool>,
    pub host_network: Option<bool>,
    pub host_pid: Option<bool>,
    pub host_users: Option<bool>,
    pub hostname: Option<String>,
    pub image_pull_secrets: Option<Vec<LocalObjectReference>>,
    pub init_containers: Option<Vec<Container>>,
    pub node_name: Option<String>,
    pub node_selector: Option<BTreeMap<String, String>>,
    pub os: Option<PodOS>,
    pub overhead: Option<BTreeMap<String, Quantity>>,
    pub preemption_policy: Option<String>,
    pub priority: Option<i32>,
    pub priority_class_name: Option<String>,
    pub readiness_gates: Option<Vec<PodReadinessGate>>,
    pub resource_claims: Option<Vec<PodResourceClaim>>,
    pub restart_policy: Option<String>,
    pub runtime_class_name: Option<String>,
    pub scheduler_name: Option<String>,
    pub scheduling_gates: Option<Vec<PodSchedulingGate>>,
    pub security_context: Option<PodSecurityContext>,
    pub service_account: Option<String>,
    pub service_account_name: Option<String>,
    pub set_hostname_as_fqdn: Option<bool>,
    pub share_process_namespace: Option<bool>,
    pub subdomain: Option<String>,
    pub termination_grace_period_seconds: Option<i64>,
    pub tolerations: Option<Vec<Toleration>>,
    pub topology_spread_constraints: Option<Vec<TopologySpreadConstraint>>,
    pub volumes: Option<Vec<Volume>>,

}

PodSpec is a description of a pod.

Fields
active_deadline_seconds: Option<i64>

Optional duration in seconds the pod may be active on the node relative to StartTime before the system will actively try to mark it failed and kill associated containers. Value must be a positive integer.

affinity: Option<Affinity>

If specified, the pod’s scheduling constraints

automount_service_account_token: Option<bool>

AutomountServiceAccountToken indicates whether a service account token should be automatically mounted.

containers: Vec<Container>

List of containers belonging to the pod. Containers cannot currently be added or removed. There must be at least one container in a Pod. Cannot be updated.

dns_config: Option<PodDNSConfig>

Specifies the DNS parameters of a pod. Parameters specified here will be merged to the generated DNS configuration based on DNSPolicy.

dns_policy: Option<String>

Set DNS policy for the pod. Defaults to “ClusterFirst”. Valid values are ‘ClusterFirstWithHostNet’, ‘ClusterFirst’, ‘Default’ or ‘None’. DNS parameters given in DNSConfig will be merged with the policy selected with DNSPolicy. To have DNS options set along with hostNetwork, you have to specify DNS policy explicitly to ‘ClusterFirstWithHostNet’.

enable_service_links: Option<bool>

EnableServiceLinks indicates whether information about services should be injected into pod’s environment variables, matching the syntax of Docker links. Optional: Defaults to true.

ephemeral_containers: Option<Vec<EphemeralContainer>>

List of ephemeral containers run in this pod. Ephemeral containers may be run in an existing pod to perform user-initiated actions such as debugging. This list cannot be specified when creating a pod, and it cannot be modified by updating the pod spec. In order to add an ephemeral container to an existing pod, use the pod’s ephemeralcontainers subresource.

host_aliases: Option<Vec<HostAlias>>

HostAliases is an optional list of hosts and IPs that will be injected into the pod’s hosts file if specified. This is only valid for non-hostNetwork pods.

host_ipc: Option<bool>

Use the host’s ipc namespace. Optional: Default to false.

host_network: Option<bool>

Host networking requested for this pod. Use the host’s network namespace. If this option is set, the ports that will be used must be specified. Default to false.

host_pid: Option<bool>

Use the host’s pid namespace. Optional: Default to false.

host_users: Option<bool>

Use the host’s user namespace. Optional: Default to true. If set to true or not present, the pod will be run in the host user namespace, useful for when the pod needs a feature only available to the host user namespace, such as loading a kernel module with CAP_SYS_MODULE. When set to false, a new userns is created for the pod. Setting false is useful for mitigating container breakout vulnerabilities even allowing users to run their containers as root without actually having root privileges on the host. This field is alpha-level and is only honored by servers that enable the UserNamespacesSupport feature.

hostname: Option<String>

Specifies the hostname of the Pod If not specified, the pod’s hostname will be set to a system-defined value.

image_pull_secrets: Option<Vec<LocalObjectReference>>

ImagePullSecrets is an optional list of references to secrets in the same namespace to use for pulling any of the images used by this PodSpec. If specified, these secrets will be passed to individual puller implementations for them to use. More info: https://kubernetes.io/docs/concepts/containers/images#specifying-imagepullsecrets-on-a-pod

init_containers: Option<Vec<Container>>

List of initialization containers belonging to the pod. Init containers are executed in order prior to containers being started. If any init container fails, the pod is considered to have failed and is handled according to its restartPolicy. The name for an init container or normal container must be unique among all containers. Init containers may not have Lifecycle actions, Readiness probes, Liveness probes, or Startup probes. The resourceRequirements of an init container are taken into account during scheduling by finding the highest request/limit for each resource type, and then using the max of of that value or the sum of the normal containers. Limits are applied to init containers in a similar fashion. Init containers cannot currently be added or removed. Cannot be updated. More info: https://kubernetes.io/docs/concepts/workloads/pods/init-containers/

node_name: Option<String>

NodeName is a request to schedule this pod onto a specific node. If it is non-empty, the scheduler simply schedules this pod onto that node, assuming that it fits resource requirements.

node_selector: Option<BTreeMap<String, String>>

NodeSelector is a selector which must be true for the pod to fit on a node. Selector which must match a node’s labels for the pod to be scheduled on that node. More info: https://kubernetes.io/docs/concepts/configuration/assign-pod-node/

os: Option<PodOS>

Specifies the OS of the containers in the pod. Some pod and container fields are restricted if this is set.

If the OS field is set to linux, the following fields must be unset: -securityContext.windowsOptions

If the OS field is set to windows, following fields must be unset: - spec.hostPID - spec.hostIPC - spec.hostUsers - spec.securityContext.seLinuxOptions - spec.securityContext.seccompProfile - spec.securityContext.fsGroup - spec.securityContext.fsGroupChangePolicy - spec.securityContext.sysctls - spec.shareProcessNamespace - spec.securityContext.runAsUser - spec.securityContext.runAsGroup - spec.securityContext.supplementalGroups - spec.containers[].securityContext.seLinuxOptions - spec.containers[].securityContext.seccompProfile - spec.containers[].securityContext.capabilities - spec.containers[].securityContext.readOnlyRootFilesystem - spec.containers[].securityContext.privileged - spec.containers[].securityContext.allowPrivilegeEscalation - spec.containers[].securityContext.procMount - spec.containers[].securityContext.runAsUser - spec.containers[*].securityContext.runAsGroup

overhead: Option<BTreeMap<String, Quantity>>

Overhead represents the resource overhead associated with running a pod for a given RuntimeClass. This field will be autopopulated at admission time by the RuntimeClass admission controller. If the RuntimeClass admission controller is enabled, overhead must not be set in Pod create requests. The RuntimeClass admission controller will reject Pod create requests which have the overhead already set. If RuntimeClass is configured and selected in the PodSpec, Overhead will be set to the value defined in the corresponding RuntimeClass, otherwise it will remain unset and treated as zero. More info: https://git.k8s.io/enhancements/keps/sig-node/688-pod-overhead/README.md

preemption_policy: Option<String>

PreemptionPolicy is the Policy for preempting pods with lower priority. One of Never, PreemptLowerPriority. Defaults to PreemptLowerPriority if unset.

priority: Option<i32>

The priority value. Various system components use this field to find the priority of the pod. When Priority Admission Controller is enabled, it prevents users from setting this field. The admission controller populates this field from PriorityClassName. The higher the value, the higher the priority.

priority_class_name: Option<String>

If specified, indicates the pod’s priority. “system-node-critical” and “system-cluster-critical” are two special keywords which indicate the highest priorities with the former being the highest priority. Any other name must be defined by creating a PriorityClass object with that name. If not specified, the pod priority will be default or zero if there is no default.

readiness_gates: Option<Vec<PodReadinessGate>>

If specified, all readiness gates will be evaluated for pod readiness. A pod is ready when all its containers are ready AND all conditions specified in the readiness gates have status equal to “True” More info: https://git.k8s.io/enhancements/keps/sig-network/580-pod-readiness-gates

resource_claims: Option<Vec<PodResourceClaim>>

ResourceClaims defines which ResourceClaims must be allocated and reserved before the Pod is allowed to start. The resources will be made available to those containers which consume them by name.

This is an alpha field and requires enabling the DynamicResourceAllocation feature gate.

This field is immutable.

restart_policy: Option<String>

Restart policy for all containers within the pod. One of Always, OnFailure, Never. In some contexts, only a subset of those values may be permitted. Default to Always. More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#restart-policy

runtime_class_name: Option<String>

RuntimeClassName refers to a RuntimeClass object in the node.k8s.io group, which should be used to run this pod. If no RuntimeClass resource matches the named class, the pod will not be run. If unset or empty, the “legacy” RuntimeClass will be used, which is an implicit class with an empty definition that uses the default runtime handler. More info: https://git.k8s.io/enhancements/keps/sig-node/585-runtime-class

scheduler_name: Option<String>

If specified, the pod will be dispatched by specified scheduler. If not specified, the pod will be dispatched by default scheduler.

scheduling_gates: Option<Vec<PodSchedulingGate>>

SchedulingGates is an opaque list of values that if specified will block scheduling the pod. If schedulingGates is not empty, the pod will stay in the SchedulingGated state and the scheduler will not attempt to schedule the pod.

SchedulingGates can only be set at pod creation time, and be removed only afterwards.

This is a beta feature enabled by the PodSchedulingReadiness feature gate.

security_context: Option<PodSecurityContext>

SecurityContext holds pod-level security attributes and common container settings. Optional: Defaults to empty. See type description for default values of each field.

service_account: Option<String>

DeprecatedServiceAccount is a depreciated alias for ServiceAccountName. Deprecated: Use serviceAccountName instead.

service_account_name: Option<String>

ServiceAccountName is the name of the ServiceAccount to use to run this pod. More info: https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/

set_hostname_as_fqdn: Option<bool>

If true the pod’s hostname will be configured as the pod’s FQDN, rather than the leaf name (the default). In Linux containers, this means setting the FQDN in the hostname field of the kernel (the nodename field of struct utsname). In Windows containers, this means setting the registry value of hostname for the registry key HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\Tcpip\Parameters to FQDN. If a pod does not have FQDN, this has no effect. Default to false.

share_process_namespace: Option<bool>

Share a single process namespace between all of the containers in a pod. When this is set containers will be able to view and signal processes from other containers in the same pod, and the first process in each container will not be assigned PID 1. HostPID and ShareProcessNamespace cannot both be set. Optional: Default to false.

subdomain: Option<String>

If specified, the fully qualified Pod hostname will be “<hostname>.<subdomain>.<pod namespace>.svc.<cluster domain>”. If not specified, the pod will not have a domainname at all.

termination_grace_period_seconds: Option<i64>

Optional duration in seconds the pod needs to terminate gracefully. May be decreased in delete request. Value must be non-negative integer. The value zero indicates stop immediately via the kill signal (no opportunity to shut down). If this value is nil, the default grace period will be used instead. The grace period is the duration in seconds after the processes running in the pod are sent a termination signal and the time when the processes are forcibly halted with a kill signal. Set this value longer than the expected cleanup time for your process. Defaults to 30 seconds.

tolerations: Option<Vec<Toleration>>

If specified, the pod’s tolerations.

topology_spread_constraints: Option<Vec<TopologySpreadConstraint>>

TopologySpreadConstraints describes how a group of pods ought to spread across topology domains. Scheduler will schedule pods in a way which abides by the constraints. All topologySpreadConstraints are ANDed.

volumes: Option<Vec<Volume>>

List of volumes that can be mounted by containers belonging to the pod. More info: https://kubernetes.io/docs/concepts/storage/volumes