Struct k8s_openapi::api::core::v1::Probe

pub struct Probe {
    pub exec: Option<ExecAction>,
    pub failure_threshold: Option<i32>,
    pub grpc: Option<GRPCAction>,
    pub http_get: Option<HTTPGetAction>,
    pub initial_delay_seconds: Option<i32>,
    pub period_seconds: Option<i32>,
    pub success_threshold: Option<i32>,
    pub tcp_socket: Option<TCPSocketAction>,
    pub termination_grace_period_seconds: Option<i64>,
    pub timeout_seconds: Option<i32>,
}

Probe describes a health check to be performed against a container to determine whether it is alive or ready to receive traffic.

Fields
exec: Option<ExecAction>

Exec specifies the action to take.

failure_threshold: Option<i32>

Minimum consecutive failures for the probe to be considered failed after having succeeded. Defaults to 3. Minimum value is 1.

grpc: Option<GRPCAction>

GRPC specifies an action involving a GRPC port. This is a beta field and requires enabling GRPCContainerProbe feature gate.

http_get: Option<HTTPGetAction>

HTTPGet specifies the http request to perform.

initial_delay_seconds: Option<i32>

Number of seconds after the container has started before liveness probes are initiated. More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle#container-probes

period_seconds: Option<i32>

How often (in seconds) to perform the probe. Default to 10 seconds. Minimum value is 1.

success_threshold: Option<i32>

Minimum consecutive successes for the probe to be considered successful after having failed. Defaults to 1. Must be 1 for liveness and startup. Minimum value is 1.

tcp_socket: Option<TCPSocketAction>

TCPSocket specifies an action involving a TCP port.

termination_grace_period_seconds: Option<i64>

Optional duration in seconds the pod needs to terminate gracefully upon probe failure. The grace period is the duration in seconds after the processes running in the pod are sent a termination signal and the time when the processes are forcibly halted with a kill signal. Set this value longer than the expected cleanup time for your process. If this value is nil, the pod’s terminationGracePeriodSeconds will be used. Otherwise, this value overrides the value provided by the pod spec. Value must be non-negative integer. The value zero indicates stop immediately via the kill signal (no opportunity to shut down). This is a beta field and requires enabling ProbeTerminationGracePeriod feature gate. Minimum value is 1. spec.terminationGracePeriodSeconds is used if unset.

timeout_seconds: Option<i32>

Number of seconds after which the probe times out. Defaults to 1 second. Minimum value is 1. More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle#container-probes