Struct k8s_openapi::api::core::v1::Service

pub struct Service {
    pub metadata: ObjectMeta,
    pub spec: Option<ServiceSpec>,
    pub status: Option<ServiceStatus>,
}

Service is a named abstraction of software service (for example, mysql) consisting of local port (for example 3306) that the proxy listens on, and the selector that determines which pods will answer requests sent through the proxy.

Fields
metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<ServiceSpec>

Spec defines the behavior of a service. https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

status: Option<ServiceStatus>

Most recently observed status of the service. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

Implementations

impl Service

pub fn connect_delete_namespaced_service_proxy(
name: &str,
namespace: &str,
optional: ConnectDeleteNamespacedServiceProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_delete_namespaced_service_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectDeleteNamespacedServiceProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_get_namespaced_service_proxy(
name: &str,
namespace: &str,
optional: ConnectGetNamespacedServiceProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_get_namespaced_service_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectGetNamespacedServiceProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_patch_namespaced_service_proxy(
name: &str,
namespace: &str,
optional: ConnectPatchNamespacedServiceProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_patch_namespaced_service_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPatchNamespacedServiceProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_post_namespaced_service_proxy(
name: &str,
namespace: &str,
optional: ConnectPostNamespacedServiceProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_post_namespaced_service_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPostNamespacedServiceProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_put_namespaced_service_proxy(
name: &str,
namespace: &str,
optional: ConnectPutNamespacedServiceProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn connect_put_namespaced_service_proxy_with_path(
name: &str,
namespace: &str,
path: &str,
optional: ConnectPutNamespacedServiceProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Service

Arguments

name

name of the ServiceProxyOptions

namespace

object name and auth scope, such as for teams and projects

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn create_namespaced_service(
namespace: &str,
body: &Service,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a Service

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn delete_collection_namespaced_service(
namespace: &str,
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of Service

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Service

pub fn delete_namespaced_service(
name: &str,
namespace: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a Service

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn list_namespaced_service(
namespace: &str,
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Service

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn list_service_for_all_namespaces(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Service

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn patch_namespaced_service(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified Service

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn patch_namespaced_service_status(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified Service

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn read_namespaced_service(
name: &str,
namespace: &str,
optional: ReadNamespacedServiceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedServiceResponse>), RequestError>

read the specified Service

Use the returned crate::ResponseBody<ReadNamespacedServiceResponse> constructor, or ReadNamespacedServiceResponse directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn read_namespaced_service_status(
name: &str,
namespace: &str,
optional: ReadNamespacedServiceStatusOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedServiceStatusResponse>), RequestError>

read status of the specified Service

Use the returned crate::ResponseBody<ReadNamespacedServiceStatusResponse> constructor, or ReadNamespacedServiceStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn replace_namespaced_service(
name: &str,
namespace: &str,
body: &Service,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified Service

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn replace_namespaced_service_status(
name: &str,
namespace: &str,
body: &Service,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified Service

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Service

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn watch_namespaced_service(
namespace: &str,
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Service

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Service

pub fn watch_service_for_all_namespaces(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Service

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.