Struct k8s_openapi::api::core::v1::ServiceAccount

pub struct ServiceAccount {
    pub automount_service_account_token: Option<bool>,
    pub image_pull_secrets: Option<Vec<LocalObjectReference>>,
    pub metadata: Option<ObjectMeta>,
    pub secrets: Option<Vec<ObjectReference>>,
}

ServiceAccount binds together: * a name, understood by users, and perhaps by peripheral systems, for an identity * a principal that can be authenticated and authorized * a set of secrets

Fields
automount_service_account_token: Option<bool>

AutomountServiceAccountToken indicates whether pods running as this service account should have an API token automatically mounted. Can be overridden at the pod level.

image_pull_secrets: Option<Vec<LocalObjectReference>>

ImagePullSecrets is a list of references to secrets in the same namespace to use for pulling any images in pods that reference this ServiceAccount. ImagePullSecrets are distinct from Secrets because Secrets can be mounted in the pod, but ImagePullSecrets are only accessed by the kubelet. More info: https://kubernetes.io/docs/concepts/containers/images/#specifying-imagepullsecrets-on-a-pod

metadata: Option<ObjectMeta>

Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

secrets: Option<Vec<ObjectReference>>

Secrets is the list of secrets allowed to be used by pods running using this ServiceAccount. More info: https://kubernetes.io/docs/concepts/configuration/secret

Methods

impl ServiceAccount

pub fn create_namespaced_service_account(
namespace: &str,
body: &ServiceAccount,
optional: CreateNamespacedServiceAccountOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateNamespacedServiceAccountResponse>), RequestError>

create a ServiceAccount

Use the returned crate::ResponseBody<CreateNamespacedServiceAccountResponse> constructor, or CreateNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn delete_collection_namespaced_service_account(
namespace: &str,
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteCollectionNamespacedServiceAccountResponse>), RequestError>

delete collection of ServiceAccount

Use the returned crate::ResponseBody<DeleteCollectionNamespacedServiceAccountResponse> constructor, or DeleteCollectionNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl ServiceAccount

pub fn delete_namespaced_service_account(
name: &str,
namespace: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteNamespacedServiceAccountResponse>), RequestError>

delete a ServiceAccount

Use the returned crate::ResponseBody<DeleteNamespacedServiceAccountResponse> constructor, or DeleteNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

name

name of the ServiceAccount

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn list_namespaced_service_account(
namespace: &str,
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListNamespacedServiceAccountResponse>), RequestError>

list or watch objects of kind ServiceAccount

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListNamespacedServiceAccountResponse> constructor, or ListNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn list_service_account_for_all_namespaces(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListServiceAccountForAllNamespacesResponse>), RequestError>

list or watch objects of kind ServiceAccount

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListServiceAccountForAllNamespacesResponse> constructor, or ListServiceAccountForAllNamespacesResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn patch_namespaced_service_account(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedServiceAccountResponse>), RequestError>

partially update the specified ServiceAccount

Use the returned crate::ResponseBody<PatchNamespacedServiceAccountResponse> constructor, or PatchNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

name

name of the ServiceAccount

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn read_namespaced_service_account(
name: &str,
namespace: &str,
optional: ReadNamespacedServiceAccountOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedServiceAccountResponse>), RequestError>

read the specified ServiceAccount

Use the returned crate::ResponseBody<ReadNamespacedServiceAccountResponse> constructor, or ReadNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

name

name of the ServiceAccount

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn replace_namespaced_service_account(
name: &str,
namespace: &str,
body: &ServiceAccount,
optional: ReplaceNamespacedServiceAccountOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedServiceAccountResponse>), RequestError>

replace the specified ServiceAccount

Use the returned crate::ResponseBody<ReplaceNamespacedServiceAccountResponse> constructor, or ReplaceNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

name

name of the ServiceAccount

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn watch_namespaced_service_account(
namespace: &str,
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchNamespacedServiceAccountResponse>), RequestError>

list or watch objects of kind ServiceAccount

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchNamespacedServiceAccountResponse> constructor, or WatchNamespacedServiceAccountResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl ServiceAccount

pub fn watch_service_account_for_all_namespaces(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchServiceAccountForAllNamespacesResponse>), RequestError>

list or watch objects of kind ServiceAccount

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchServiceAccountForAllNamespacesResponse> constructor, or WatchServiceAccountForAllNamespacesResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.