Struct k8s_openapi::api::discovery::v1::Endpoint

pub struct Endpoint {
    pub addresses: Vec<String>,
    pub conditions: Option<EndpointConditions>,
    pub deprecated_topology: Option<BTreeMap<String, String>>,
    pub hints: Option<EndpointHints>,
    pub hostname: Option<String>,
    pub node_name: Option<String>,
    pub target_ref: Option<ObjectReference>,
    pub zone: Option<String>,
}

Endpoint represents a single logical “backend” implementing a service.

Fields
addresses: Vec<String>

addresses of this endpoint. The contents of this field are interpreted according to the corresponding EndpointSlice addressType field. Consumers must handle different types of addresses in the context of their own capabilities. This must contain at least one address but no more than 100. These are all assumed to be fungible and clients may choose to only use the first element. Refer to: https://issue.k8s.io/106267

conditions: Option<EndpointConditions>

conditions contains information about the current status of the endpoint.

deprecated_topology: Option<BTreeMap<String, String>>

deprecatedTopology contains topology information part of the v1beta1 API. This field is deprecated, and will be removed when the v1beta1 API is removed (no sooner than kubernetes v1.24). While this field can hold values, it is not writable through the v1 API, and any attempts to write to it will be silently ignored. Topology information can be found in the zone and nodeName fields instead.

hints: Option<EndpointHints>

hints contains information associated with how an endpoint should be consumed.

hostname: Option<String>

hostname of this endpoint. This field may be used by consumers of endpoints to distinguish endpoints from each other (e.g. in DNS names). Multiple endpoints which use the same hostname should be considered fungible (e.g. multiple A values in DNS). Must be lowercase and pass DNS Label (RFC 1123) validation.

node_name: Option<String>

nodeName represents the name of the Node hosting this endpoint. This can be used to determine endpoints local to a Node. This field can be enabled with the EndpointSliceNodeName feature gate.

target_ref: Option<ObjectReference>

targetRef is a reference to a Kubernetes object that represents this endpoint.

zone: Option<String>

zone is the name of the Zone this endpoint exists in.