Module k8s_openapi::api::discovery::v1alpha1

Structs

CreateNamespacedEndpointSliceOptional
Optional parameters of EndpointSlice::create_namespaced_endpoint_slice

Endpoint
Endpoint represents a single logical "backend" implementing a service.

EndpointConditions
EndpointConditions represents the current condition of an endpoint.

EndpointPort
EndpointPort represents a Port used by an EndpointSlice

EndpointSlice
EndpointSlice represents a subset of the endpoints that implement a service. For a given service there may be multiple EndpointSlice objects, selected by labels, which must be joined to produce the full set of endpoints.

EndpointSliceList
EndpointSliceList represents a list of endpoint slices

ReadNamespacedEndpointSliceOptional
Optional parameters of EndpointSlice::read_namespaced_endpoint_slice

ReplaceNamespacedEndpointSliceOptional
Optional parameters of EndpointSlice::replace_namespaced_endpoint_slice

Enums

CreateNamespacedEndpointSliceResponse
Use <CreateNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::create_namespaced_endpoint_slice

DeleteCollectionNamespacedEndpointSliceResponse
Use <DeleteCollectionNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::delete_collection_namespaced_endpoint_slice

DeleteNamespacedEndpointSliceResponse
Use <DeleteNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::delete_namespaced_endpoint_slice

ListEndpointSliceForAllNamespacesResponse
Use <ListEndpointSliceForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::list_endpoint_slice_for_all_namespaces

ListNamespacedEndpointSliceResponse
Use <ListNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::list_namespaced_endpoint_slice

PatchNamespacedEndpointSliceResponse
Use <PatchNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::patch_namespaced_endpoint_slice

ReadNamespacedEndpointSliceResponse
Use <ReadNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::read_namespaced_endpoint_slice

ReplaceNamespacedEndpointSliceResponse
Use <ReplaceNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::replace_namespaced_endpoint_slice

WatchEndpointSliceForAllNamespacesResponse
Use <WatchEndpointSliceForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::watch_endpoint_slice_for_all_namespaces

WatchNamespacedEndpointSliceResponse
Use <WatchNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::watch_namespaced_endpoint_slice