Struct k8s_openapi::api::events::v1::Event

pub struct Event {

    pub action: Option<String>,
    pub deprecated_count: Option<i32>,
    pub deprecated_first_timestamp: Option<Time>,
    pub deprecated_last_timestamp: Option<Time>,
    pub deprecated_source: Option<EventSource>,
    pub event_time: Option<MicroTime>,
    pub metadata: ObjectMeta,
    pub note: Option<String>,
    pub reason: Option<String>,
    pub regarding: Option<ObjectReference>,
    pub related: Option<ObjectReference>,
    pub reporting_controller: Option<String>,
    pub reporting_instance: Option<String>,
    pub series: Option<EventSeries>,
    pub type_: Option<String>,

}

Event is a report of an event somewhere in the cluster. It generally denotes some state change in the system. Events have a limited retention time and triggers and messages may evolve with time. Event consumers should not rely on the timing of an event with a given Reason reflecting a consistent underlying trigger, or the continued existence of events with that Reason. Events should be treated as informative, best-effort, supplemental data.

Fields
action: Option<String>

action is what action was taken/failed regarding to the regarding object. It is machine-readable. This field cannot be empty for new Events and it can have at most 128 characters.

deprecated_count: Option<i32>

deprecatedCount is the deprecated field assuring backward compatibility with core.v1 Event type.

deprecated_first_timestamp: Option<Time>

deprecatedFirstTimestamp is the deprecated field assuring backward compatibility with core.v1 Event type.

deprecated_last_timestamp: Option<Time>

deprecatedLastTimestamp is the deprecated field assuring backward compatibility with core.v1 Event type.

deprecated_source: Option<EventSource>

deprecatedSource is the deprecated field assuring backward compatibility with core.v1 Event type.

event_time: Option<MicroTime>

eventTime is the time when this Event was first observed. It is required.

metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

note: Option<String>

note is a human-readable description of the status of this operation. Maximal length of the note is 1kB, but libraries should be prepared to handle values up to 64kB.

reason: Option<String>

reason is why the action was taken. It is human-readable. This field cannot be empty for new Events and it can have at most 128 characters.

regarding: Option<ObjectReference>

regarding contains the object this Event is about. In most cases it’s an Object reporting controller implements, e.g. ReplicaSetController implements ReplicaSets and this event is emitted because it acts on some changes in a ReplicaSet object.

related: Option<ObjectReference>

related is the optional secondary object for more complex actions. E.g. when regarding object triggers a creation or deletion of related object.

reporting_controller: Option<String>

reportingController is the name of the controller that emitted this Event, e.g. kubernetes.io/kubelet. This field cannot be empty for new Events.

reporting_instance: Option<String>

reportingInstance is the ID of the controller instance, e.g. kubelet-xyzf. This field cannot be empty for new Events and it can have at most 128 characters.

series: Option<EventSeries>

series is data about the Event series this event represents or nil if it’s a singleton Event.

type_: Option<String>

type is the type of this event (Normal, Warning), new types could be added in the future. It is machine-readable. This field cannot be empty for new Events.