Module k8s_openapi::api::events::v1beta1

Structs

Event
Event is a report of an event somewhere in the cluster. It generally denotes some state change in the system. Events have a limited retention time and triggers and messages may evolve with time. Event consumers should not rely on the timing of an event with a given Reason reflecting a consistent underlying trigger, or the continued existence of events with that Reason. Events should be treated as informative, best-effort, supplemental data.

EventSeries
EventSeries contain information on series of events, i.e. thing that was/is happening continuously for some time.

ReadNamespacedEventOptional
Optional parameters of Event::read_namespaced_event

Enums

ReadNamespacedEventResponse
Use <ReadNamespacedEventResponse as Response>::try_from_parts to parse the HTTP response body of Event::read_namespaced_event