Module k8s_openapi::api::node::v1beta1

Structs

Overhead
Overhead structure represents the resource overhead associated with running a pod.

ReadRuntimeClassOptional
Optional parameters of RuntimeClass::read_runtime_class

RuntimeClass
RuntimeClass defines a class of container runtime supported in the cluster. The RuntimeClass is used to determine which container runtime is used to run all containers in a pod. RuntimeClasses are (currently) manually defined by a user or cluster provisioner, and referenced in the PodSpec. The Kubelet is responsible for resolving the RuntimeClassName reference before running the pod. For more details, see https://git.k8s.io/enhancements/keps/sig-node/runtime-class.md

Scheduling
Scheduling specifies the scheduling constraints for nodes supporting a RuntimeClass.

Enums

ReadRuntimeClassResponse
Use <ReadRuntimeClassResponse as Response>::try_from_parts to parse the HTTP response body of RuntimeClass::read_runtime_class