Struct k8s_openapi::api::node::v1beta1::RuntimeClass

pub struct RuntimeClass {
    pub handler: String,
    pub metadata: Option<ObjectMeta>,
    pub overhead: Option<Overhead>,
    pub scheduling: Option<Scheduling>,
}

RuntimeClass defines a class of container runtime supported in the cluster. The RuntimeClass is used to determine which container runtime is used to run all containers in a pod. RuntimeClasses are (currently) manually defined by a user or cluster provisioner, and referenced in the PodSpec. The Kubelet is responsible for resolving the RuntimeClassName reference before running the pod. For more details, see https://git.k8s.io/enhancements/keps/sig-node/runtime-class.md

Fields
handler: String

Handler specifies the underlying runtime and configuration that the CRI implementation will use to handle pods of this class. The possible values are specific to the node & CRI configuration. It is assumed that all handlers are available on every node, and handlers of the same name are equivalent on every node. For example, a handler called "runc" might specify that the runc OCI runtime (using native Linux containers) will be used to run the containers in a pod. The Handler must conform to the DNS Label (RFC 1123) requirements, and is immutable.

metadata: Option<ObjectMeta>

More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

overhead: Option<Overhead>

Overhead represents the resource overhead associated with running a pod for a given RuntimeClass. For more details, see https://git.k8s.io/enhancements/keps/sig-node/20190226-pod-overhead.md This field is alpha-level as of Kubernetes v1.15, and is only honored by servers that enable the PodOverhead feature.

scheduling: Option<Scheduling>

Scheduling holds the scheduling constraints to ensure that pods running with this RuntimeClass are scheduled to nodes that support it. If scheduling is nil, this RuntimeClass is assumed to be supported by all nodes.

Methods

impl RuntimeClass

pub fn create_runtime_class(
body: &RuntimeClass,
optional: CreateRuntimeClassOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateRuntimeClassResponse>), RequestError>

create a RuntimeClass

Use the returned crate::ResponseBody<CreateRuntimeClassResponse> constructor, or CreateRuntimeClassResponse directly, to parse the HTTP response.

Arguments

body

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn delete_collection_runtime_class(
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteCollectionRuntimeClassResponse>), RequestError>

delete collection of RuntimeClass

Use the returned crate::ResponseBody<DeleteCollectionRuntimeClassResponse> constructor, or DeleteCollectionRuntimeClassResponse directly, to parse the HTTP response.

Arguments

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl RuntimeClass

pub fn delete_runtime_class(
name: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteRuntimeClassResponse>), RequestError>

delete a RuntimeClass

Use the returned crate::ResponseBody<DeleteRuntimeClassResponse> constructor, or DeleteRuntimeClassResponse directly, to parse the HTTP response.

Arguments

name

name of the RuntimeClass

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn list_runtime_class(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListRuntimeClassResponse>), RequestError>

list or watch objects of kind RuntimeClass

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListRuntimeClassResponse> constructor, or ListRuntimeClassResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn patch_runtime_class(
name: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchRuntimeClassResponse>), RequestError>

partially update the specified RuntimeClass

Use the returned crate::ResponseBody<PatchRuntimeClassResponse> constructor, or PatchRuntimeClassResponse directly, to parse the HTTP response.

Arguments

name

name of the RuntimeClass

body

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn read_runtime_class(
name: &str,
optional: ReadRuntimeClassOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadRuntimeClassResponse>), RequestError>

read the specified RuntimeClass

Use the returned crate::ResponseBody<ReadRuntimeClassResponse> constructor, or ReadRuntimeClassResponse directly, to parse the HTTP response.

Arguments

name

name of the RuntimeClass

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn replace_runtime_class(
name: &str,
body: &RuntimeClass,
optional: ReplaceRuntimeClassOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceRuntimeClassResponse>), RequestError>

replace the specified RuntimeClass

Use the returned crate::ResponseBody<ReplaceRuntimeClassResponse> constructor, or ReplaceRuntimeClassResponse directly, to parse the HTTP response.

Arguments

name

name of the RuntimeClass

body

optional

Optional parameters. Use Default::default() to not pass any.

impl RuntimeClass

pub fn watch_runtime_class(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchRuntimeClassResponse>), RequestError>

list or watch objects of kind RuntimeClass

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchRuntimeClassResponse> constructor, or WatchRuntimeClassResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.