Module k8s_openapi::api::policy::v1

Structs

Eviction

Eviction evicts a pod from its node subject to certain policies and safety constraints. This is a subresource of Pod. A request to cause such an eviction is created by POSTing to …/pods/<pod name>/evictions.

PodDisruptionBudget

PodDisruptionBudget is an object to define the max disruption that can be caused to a collection of pods

PodDisruptionBudgetSpec

PodDisruptionBudgetSpec is a description of a PodDisruptionBudget.

PodDisruptionBudgetStatus

PodDisruptionBudgetStatus represents information about the status of a PodDisruptionBudget. Status may trail the actual state of a system.

Enums

ReadPodDisruptionBudgetResponse

Use <ReadPodDisruptionBudgetResponse as Response>::try_from_parts to parse the HTTP response body of PodDisruptionBudget::read

ReadPodDisruptionBudgetStatusResponse

Use <ReadPodDisruptionBudgetStatusResponse as Response>::try_from_parts to parse the HTTP response body of PodDisruptionBudget::read_status