Struct k8s_openapi::api::policy::v1beta1::PodDisruptionBudget

pub struct PodDisruptionBudget {
    pub metadata: Option<ObjectMeta>,
    pub spec: Option<PodDisruptionBudgetSpec>,
    pub status: Option<PodDisruptionBudgetStatus>,
}

PodDisruptionBudget is an object to define the max disruption that can be caused to a collection of pods

Fields
metadata: Option<ObjectMeta>spec: Option<PodDisruptionBudgetSpec>

Specification of the desired behavior of the PodDisruptionBudget.

status: Option<PodDisruptionBudgetStatus>

Most recently observed status of the PodDisruptionBudget.

Methods

impl PodDisruptionBudget

pub fn create_namespaced_pod_disruption_budget(
namespace: &str,
body: &PodDisruptionBudget,
optional: CreateOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a PodDisruptionBudget

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn delete_collection_namespaced_pod_disruption_budget(
namespace: &str,
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of PodDisruptionBudget

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn delete_namespaced_pod_disruption_budget(
name: &str,
namespace: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a PodDisruptionBudget

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn list_namespaced_pod_disruption_budget(
namespace: &str,
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind PodDisruptionBudget

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or [crate::ListResponse]`` directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn list_pod_disruption_budget_for_all_namespaces(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind PodDisruptionBudget

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or [crate::ListResponse]`` directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn patch_namespaced_pod_disruption_budget(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified PodDisruptionBudget

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn patch_namespaced_pod_disruption_budget_status(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified PodDisruptionBudget

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn read_namespaced_pod_disruption_budget(
name: &str,
namespace: &str,
optional: ReadNamespacedPodDisruptionBudgetOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedPodDisruptionBudgetResponse>), RequestError>

read the specified PodDisruptionBudget

Use the returned crate::ResponseBody<ReadNamespacedPodDisruptionBudgetResponse> constructor, or ReadNamespacedPodDisruptionBudgetResponse directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn read_namespaced_pod_disruption_budget_status(
name: &str,
namespace: &str,
optional: ReadNamespacedPodDisruptionBudgetStatusOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedPodDisruptionBudgetStatusResponse>), RequestError>

read status of the specified PodDisruptionBudget

Use the returned crate::ResponseBody<ReadNamespacedPodDisruptionBudgetStatusResponse> constructor, or ReadNamespacedPodDisruptionBudgetStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn replace_namespaced_pod_disruption_budget(
name: &str,
namespace: &str,
body: &PodDisruptionBudget,
optional: ReplaceOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified PodDisruptionBudget

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn replace_namespaced_pod_disruption_budget_status(
name: &str,
namespace: &str,
body: &PodDisruptionBudget,
optional: ReplaceOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified PodDisruptionBudget

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the PodDisruptionBudget

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn watch_namespaced_pod_disruption_budget(
namespace: &str,
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind PodDisruptionBudget

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl PodDisruptionBudget

pub fn watch_pod_disruption_budget_for_all_namespaces(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind PodDisruptionBudget

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.