Struct k8s_openapi::api::policy::v1beta1::PodSecurityPolicySpec

pub struct PodSecurityPolicySpec {

    pub allow_privilege_escalation: Option<bool>,
    pub allowed_csi_drivers: Option<Vec<AllowedCSIDriver>>,
    pub allowed_capabilities: Option<Vec<String>>,
    pub allowed_flex_volumes: Option<Vec<AllowedFlexVolume>>,
    pub allowed_host_paths: Option<Vec<AllowedHostPath>>,
    pub allowed_proc_mount_types: Option<Vec<String>>,
    pub allowed_unsafe_sysctls: Option<Vec<String>>,
    pub default_add_capabilities: Option<Vec<String>>,
    pub default_allow_privilege_escalation: Option<bool>,
    pub forbidden_sysctls: Option<Vec<String>>,
    pub fs_group: FSGroupStrategyOptions,
    pub host_ipc: Option<bool>,
    pub host_network: Option<bool>,
    pub host_pid: Option<bool>,
    pub host_ports: Option<Vec<HostPortRange>>,
    pub privileged: Option<bool>,
    pub read_only_root_filesystem: Option<bool>,
    pub required_drop_capabilities: Option<Vec<String>>,
    pub run_as_group: Option<RunAsGroupStrategyOptions>,
    pub run_as_user: RunAsUserStrategyOptions,
    pub runtime_class: Option<RuntimeClassStrategyOptions>,
    pub se_linux: SELinuxStrategyOptions,
    pub supplemental_groups: SupplementalGroupsStrategyOptions,
    pub volumes: Option<Vec<String>>,

}

PodSecurityPolicySpec defines the policy enforced.

Fields
allow_privilege_escalation: Option<bool>

allowPrivilegeEscalation determines if a pod can request to allow privilege escalation. If unspecified, defaults to true.

allowed_csi_drivers: Option<Vec<AllowedCSIDriver>>

AllowedCSIDrivers is an allowlist of inline CSI drivers that must be explicitly set to be embedded within a pod spec. An empty value indicates that any CSI driver can be used for inline ephemeral volumes. This is a beta field, and is only honored if the API server enables the CSIInlineVolume feature gate.

allowed_capabilities: Option<Vec<String>>

allowedCapabilities is a list of capabilities that can be requested to add to the container. Capabilities in this field may be added at the pod author’s discretion. You must not list a capability in both allowedCapabilities and requiredDropCapabilities.

allowed_flex_volumes: Option<Vec<AllowedFlexVolume>>

allowedFlexVolumes is an allowlist of Flexvolumes. Empty or nil indicates that all Flexvolumes may be used. This parameter is effective only when the usage of the Flexvolumes is allowed in the “volumes” field.

allowed_host_paths: Option<Vec<AllowedHostPath>>

allowedHostPaths is an allowlist of host paths. Empty indicates that all host paths may be used.

allowed_proc_mount_types: Option<Vec<String>>

AllowedProcMountTypes is an allowlist of allowed ProcMountTypes. Empty or nil indicates that only the DefaultProcMountType may be used. This requires the ProcMountType feature flag to be enabled.

allowed_unsafe_sysctls: Option<Vec<String>>

allowedUnsafeSysctls is a list of explicitly allowed unsafe sysctls, defaults to none. Each entry is either a plain sysctl name or ends in “*” in which case it is considered as a prefix of allowed sysctls. Single * means all unsafe sysctls are allowed. Kubelet has to allowlist all allowed unsafe sysctls explicitly to avoid rejection.

Examples: e.g. “foo/” allows “foo/bar”, “foo/baz”, etc. e.g. “foo.” allows “foo.bar”, “foo.baz”, etc.

default_add_capabilities: Option<Vec<String>>

defaultAddCapabilities is the default set of capabilities that will be added to the container unless the pod spec specifically drops the capability. You may not list a capability in both defaultAddCapabilities and requiredDropCapabilities. Capabilities added here are implicitly allowed, and need not be included in the allowedCapabilities list.

default_allow_privilege_escalation: Option<bool>

defaultAllowPrivilegeEscalation controls the default setting for whether a process can gain more privileges than its parent process.

forbidden_sysctls: Option<Vec<String>>

forbiddenSysctls is a list of explicitly forbidden sysctls, defaults to none. Each entry is either a plain sysctl name or ends in “*” in which case it is considered as a prefix of forbidden sysctls. Single * means all sysctls are forbidden.

Examples: e.g. “foo/” forbids “foo/bar”, “foo/baz”, etc. e.g. “foo.” forbids “foo.bar”, “foo.baz”, etc.

fs_group: FSGroupStrategyOptions

fsGroup is the strategy that will dictate what fs group is used by the SecurityContext.

host_ipc: Option<bool>

hostIPC determines if the policy allows the use of HostIPC in the pod spec.

host_network: Option<bool>

hostNetwork determines if the policy allows the use of HostNetwork in the pod spec.

host_pid: Option<bool>

hostPID determines if the policy allows the use of HostPID in the pod spec.

host_ports: Option<Vec<HostPortRange>>

hostPorts determines which host port ranges are allowed to be exposed.

privileged: Option<bool>

privileged determines if a pod can request to be run as privileged.

read_only_root_filesystem: Option<bool>

readOnlyRootFilesystem when set to true will force containers to run with a read only root file system. If the container specifically requests to run with a non-read only root file system the PSP should deny the pod. If set to false the container may run with a read only root file system if it wishes but it will not be forced to.

required_drop_capabilities: Option<Vec<String>>

requiredDropCapabilities are the capabilities that will be dropped from the container. These are required to be dropped and cannot be added.

run_as_group: Option<RunAsGroupStrategyOptions>

RunAsGroup is the strategy that will dictate the allowable RunAsGroup values that may be set. If this field is omitted, the pod’s RunAsGroup can take any value. This field requires the RunAsGroup feature gate to be enabled.

run_as_user: RunAsUserStrategyOptions

runAsUser is the strategy that will dictate the allowable RunAsUser values that may be set.

runtime_class: Option<RuntimeClassStrategyOptions>

runtimeClass is the strategy that will dictate the allowable RuntimeClasses for a pod. If this field is omitted, the pod’s runtimeClassName field is unrestricted. Enforcement of this field depends on the RuntimeClass feature gate being enabled.

se_linux: SELinuxStrategyOptions

seLinux is the strategy that will dictate the allowable labels that may be set.

supplemental_groups: SupplementalGroupsStrategyOptions

supplementalGroups is the strategy that will dictate what supplemental groups are used by the SecurityContext.

volumes: Option<Vec<String>>

volumes is an allowlist of volume plugins. Empty indicates that no volumes may be used. To allow all volumes you may use ‘*’.