k8s_openapi::api::resource

Module v1

Structs

AllocatedDeviceStatus

AllocatedDeviceStatus contains the status of an allocated device, if the driver chooses to report it. This may include driver-specific information.

AllocationResult

AllocationResult contains attributes of an allocated resource.

CELDeviceSelector

CELDeviceSelector contains a CEL expression for selecting a device.

CapacityRequestPolicy

CapacityRequestPolicy defines how requests consume device capacity.

CapacityRequestPolicyRange

CapacityRequestPolicyRange defines a valid range for consumable capacity values.

CapacityRequirements

CapacityRequirements defines the capacity requirements for a specific device request.

Counter

Counter describes a quantity associated with a device.

CounterSet

CounterSet defines a named set of counters that are available to be used by devices defined in the ResourceSlice.

Device

Device represents one individual hardware instance that can be selected based on its attributes. Besides the name, exactly one field must be set.

DeviceAllocationConfiguration

DeviceAllocationConfiguration gets embedded in an AllocationResult.

DeviceAllocationResult

DeviceAllocationResult is the result of allocating devices.

DeviceAttribute

DeviceAttribute must have exactly one field set.

DeviceCapacity

DeviceCapacity describes a quantity associated with a device.

DeviceClaim

DeviceClaim defines how to request devices with a ResourceClaim.

DeviceClaimConfiguration

DeviceClaimConfiguration is used for configuration parameters in DeviceClaim.

DeviceClass

DeviceClass is a vendor- or admin-provided resource that contains device configuration and selectors. It can be referenced in the device requests of a claim to apply these presets. Cluster scoped.

DeviceClassConfiguration

DeviceClassConfiguration is used in DeviceClass.

DeviceClassSpec

DeviceClassSpec is used in a [DeviceClass] to define what can be allocated and how to configure it.

DeviceConstraint

DeviceConstraint must have exactly one field set besides Requests.

DeviceCounterConsumption

DeviceCounterConsumption defines a set of counters that a device will consume from a CounterSet.

DeviceRequest

DeviceRequest is a request for devices required for a claim. This is typically a request for a single resource like a device, but can also ask for several identical devices. With FirstAvailable it is also possible to provide a prioritized list of requests.

DeviceRequestAllocationResult

DeviceRequestAllocationResult contains the allocation result for one request.

DeviceSelector

DeviceSelector must have exactly one field set.

DeviceSubRequest

DeviceSubRequest describes a request for device provided in the claim.spec.devices.requests[].firstAvailable array. Each is typically a request for a single resource like a device, but can also ask for several identical devices.

DeviceTaint

The device this taint is attached to has the “effect” on any claim which does not tolerate the taint and, through the claim, to pods using the claim.

DeviceToleration

The ResourceClaim this DeviceToleration is attached to tolerates any taint that matches the triple <key,value,effect> using the matching operator <operator>.

ExactDeviceRequest

ExactDeviceRequest is a request for one or more identical devices.

NetworkDeviceData

NetworkDeviceData provides network-related details for the allocated device. This information may be filled by drivers or other components to configure or identify the device within a network context.

OpaqueDeviceConfiguration

OpaqueDeviceConfiguration contains configuration parameters for a driver in a format defined by the driver vendor.

ResourceClaim

ResourceClaim describes a request for access to resources in the cluster, for use by workloads. For example, if a workload needs an accelerator device with specific properties, this is how that request is expressed. The status stanza tracks whether this claim has been satisfied and what specific resources have been allocated.

ResourceClaimConsumerReference

ResourceClaimConsumerReference contains enough information to let you locate the consumer of a ResourceClaim. The user must be a resource in the same namespace as the ResourceClaim.

ResourceClaimSpec

ResourceClaimSpec defines what is being requested in a ResourceClaim and how to configure it.

ResourceClaimStatus

ResourceClaimStatus tracks whether the resource has been allocated and what the result of that was.

ResourceClaimTemplate

ResourceClaimTemplate is used to produce ResourceClaim objects.

ResourceClaimTemplateSpec

ResourceClaimTemplateSpec contains the metadata and fields for a ResourceClaim.

ResourcePool

ResourcePool describes the pool that ResourceSlices belong to.

ResourceSlice

ResourceSlice represents one or more resources in a pool of similar resources, managed by a common driver. A pool may span more than one ResourceSlice, and exactly how many ResourceSlices comprise a pool is determined by the driver.

ResourceSliceSpec

ResourceSliceSpec contains the information published by the driver in one ResourceSlice.