Module k8s_openapi::api::resource::v1alpha1

Structs

AllocationResult

AllocationResult contains attributed of an allocated resource.

PodScheduling

PodScheduling objects hold information that is needed to schedule a Pod with ResourceClaims that use “WaitForFirstConsumer” allocation mode.

PodSchedulingSpec

PodSchedulingSpec describes where resources for the Pod are needed.

PodSchedulingStatus

PodSchedulingStatus describes where resources for the Pod can be allocated.

ResourceClaim

ResourceClaim describes which resources are needed by a resource consumer. Its status tracks whether the resource has been allocated and what the resulting attributes are.

ResourceClaimConsumerReference

ResourceClaimConsumerReference contains enough information to let you locate the consumer of a ResourceClaim. The user must be a resource in the same namespace as the ResourceClaim.

ResourceClaimParametersReference

ResourceClaimParametersReference contains enough information to let you locate the parameters for a ResourceClaim. The object must be in the same namespace as the ResourceClaim.

ResourceClaimSchedulingStatus

ResourceClaimSchedulingStatus contains information about one particular ResourceClaim with “WaitForFirstConsumer” allocation mode.

ResourceClaimSpec

ResourceClaimSpec defines how a resource is to be allocated.

ResourceClaimStatus

ResourceClaimStatus tracks whether the resource has been allocated and what the resulting attributes are.

ResourceClaimTemplate

ResourceClaimTemplate is used to produce ResourceClaim objects.

ResourceClaimTemplateSpec

ResourceClaimTemplateSpec contains the metadata and fields for a ResourceClaim.

ResourceClass

ResourceClass is used by administrators to influence how resources are allocated.

ResourceClassParametersReference

ResourceClassParametersReference contains enough information to let you locate the parameters for a ResourceClass.