k8s_openapi::api::resource

Module v1alpha2

Structs

AllocationResult

AllocationResult contains attributes of an allocated resource.

DriverAllocationResult

DriverAllocationResult contains vendor parameters and the allocation result for one request.

DriverRequests

DriverRequests describes all resources that are needed from one particular driver.

NamedResourcesAllocationResult

NamedResourcesAllocationResult is used in AllocationResultModel.

NamedResourcesAttribute

NamedResourcesAttribute is a combination of an attribute name and its value.

NamedResourcesFilter

NamedResourcesFilter is used in ResourceFilterModel.

NamedResourcesInstance

NamedResourcesInstance represents one individual hardware instance that can be selected based on its attributes.

NamedResourcesIntSlice

NamedResourcesIntSlice contains a slice of 64-bit integers.

NamedResourcesRequest

NamedResourcesRequest is used in ResourceRequestModel.

NamedResourcesResources

NamedResourcesResources is used in ResourceModel.

NamedResourcesStringSlice

NamedResourcesStringSlice contains a slice of strings.

PodSchedulingContext

PodSchedulingContext objects hold information that is needed to schedule a Pod with ResourceClaims that use “WaitForFirstConsumer” allocation mode.

PodSchedulingContextSpec

PodSchedulingContextSpec describes where resources for the Pod are needed.

PodSchedulingContextStatus

PodSchedulingContextStatus describes where resources for the Pod can be allocated.

ResourceClaim

ResourceClaim describes which resources are needed by a resource consumer. Its status tracks whether the resource has been allocated and what the resulting attributes are.

ResourceClaimConsumerReference

ResourceClaimConsumerReference contains enough information to let you locate the consumer of a ResourceClaim. The user must be a resource in the same namespace as the ResourceClaim.

ResourceClaimParameters

ResourceClaimParameters defines resource requests for a ResourceClaim in an in-tree format understood by Kubernetes.

ResourceClaimParametersReference

ResourceClaimParametersReference contains enough information to let you locate the parameters for a ResourceClaim. The object must be in the same namespace as the ResourceClaim.

ResourceClaimSchedulingStatus

ResourceClaimSchedulingStatus contains information about one particular ResourceClaim with “WaitForFirstConsumer” allocation mode.

ResourceClaimSpec

ResourceClaimSpec defines how a resource is to be allocated.

ResourceClaimStatus

ResourceClaimStatus tracks whether the resource has been allocated and what the resulting attributes are.

ResourceClaimTemplate

ResourceClaimTemplate is used to produce ResourceClaim objects.

ResourceClaimTemplateSpec

ResourceClaimTemplateSpec contains the metadata and fields for a ResourceClaim.

ResourceClass

ResourceClass is used by administrators to influence how resources are allocated.

ResourceClassParameters

ResourceClassParameters defines resource requests for a ResourceClass in an in-tree format understood by Kubernetes.

ResourceClassParametersReference

ResourceClassParametersReference contains enough information to let you locate the parameters for a ResourceClass.

ResourceFilter

ResourceFilter is a filter for resources from one particular driver.

ResourceHandle

ResourceHandle holds opaque resource data for processing by a specific kubelet plugin.

ResourceRequest

ResourceRequest is a request for resources from one particular driver.

ResourceSlice

ResourceSlice provides information about available resources on individual nodes.

StructuredResourceHandle

StructuredResourceHandle is the in-tree representation of the allocation result.

VendorParameters

VendorParameters are opaque parameters for one particular driver.