k8s_openapi::api::resource

Module v1alpha3

Structs

CELDeviceSelector

CELDeviceSelector contains a CEL expression for selecting a device.

DeviceSelector

DeviceSelector must have exactly one field set.

DeviceTaint

The device this taint is attached to has the “effect” on any claim which does not tolerate the taint and, through the claim, to pods using the claim.

DeviceTaintRule

DeviceTaintRule adds one taint to all devices which match the selector. This has the same effect as if the taint was specified directly in the ResourceSlice by the DRA driver.

DeviceTaintRuleSpec

DeviceTaintRuleSpec specifies the selector and one taint.

DeviceTaintSelector

DeviceTaintSelector defines which device(s) a DeviceTaintRule applies to. The empty selector matches all devices. Without a selector, no devices are matched.