k8s_openapi::kube_aggregator::pkg::apis::apiregistration
Module v1

Structs

APIService

APIService represents a server for a particular GroupVersion. Name must be “version.group”.

APIServiceCondition

APIServiceCondition describes the state of an APIService at a particular point

APIServiceSpec

APIServiceSpec contains information for locating and communicating with a server. Only https is supported, though you are able to disable certificate verification.

APIServiceStatus

APIServiceStatus contains derived information about an API server

ServiceReference

ServiceReference holds a reference to Service.legacy.k8s.io