Module k8s_openapi::kube_aggregator::pkg::apis::apiregistration::v1beta1

Structs

APIService
APIService represents a server for a particular GroupVersion. Name must be “version.group”.

APIServiceCondition
APIServiceCondition describes the state of an APIService at a particular point

APIServiceSpec
APIServiceSpec contains information for locating and communicating with a server. Only https is supported, though you are able to disable certificate verification.

APIServiceStatus
APIServiceStatus contains derived information about an API server

ReadAPIServiceOptional
Optional parameters of APIService::read_api_service

ReadAPIServiceStatusOptional
Optional parameters of APIService::read_api_service_status

ServiceReference
ServiceReference holds a reference to Service.legacy.k8s.io

Enums

ReadAPIServiceResponse
Use <ReadAPIServiceResponse as Response>::try_from_parts to parse the HTTP response body of APIService::read_api_service

ReadAPIServiceStatusResponse
Use <ReadAPIServiceStatusResponse as Response>::try_from_parts to parse the HTTP response body of APIService::read_api_service_status