Macro k8s_openapi::k8s_if_ge_1_18

macro_rules! k8s_if_ge_1_18 {
    ($($tt:tt)*) => { ... };
}

This macro evaluates to its contents if the v1_18 or higher feature is enabled, otherwise it evaluates to nothing.