Macro k8s_openapi::k8s_if_ge_1_26

macro_rules! k8s_if_ge_1_26 {
    ($($tt:tt)*) => { ... };
}

This macro evaluates to its contents if the v1_26 or higher feature is enabled, otherwise it evaluates to nothing.