k8s_openapi

Macro k8s_if_ge_1_31

macro_rules! k8s_if_ge_1_31 {
    ($($tt:tt)*) => { ... };
}

This macro evaluates to its contents if the v1_31 or higher feature is enabled, otherwise it evaluates to nothing.