Trait k8s_openapi::DeepMerge

pub trait DeepMerge {
    // Required method
    fn merge_from(&mut self, other: Self);
}

A trait applies to types that support deep merging.

current.merge_from(new) behaves in the following ways:

structs

Structs are merged by individually merging each of their fields. For example, given:

struct S {
    a: i32,
    b: String,
}

… the expected impl of DeepMerge for S would be:

impl DeepMerge for S {
    fn merge_from(&mut self, other: Self) {
        self.a.merge_from(other.a);
        self.b.merge_from(other.b);
    }
}

The structs in the k8s-openapi crate behave this way. If you are implementing this trait for your own types, it is recommended to impl it in the same way.

Option

If new is a None, current is unchanged.

If new is a Some(new_inner):

If current is a None, current becomes Some(new_inner).

If current is a Some(current_inner), current_inner is merged with new_inner.

Vec

Use an explicit merge strategy.

BTreeMap

Use an explicit merge strategy.

serde_json::Value

serde_json::Value is merged using the JSON merge algorithm (RFC 7396).

Other types

self is just replaced by other.

Required Methods

fn merge_from(&mut self, other: Self)

Merge other into self.

Object Safety

This trait is not object safe.