Trait k8s_openapi::ListableResource

pub trait ListableResource: Resource {
    const LIST_KIND: &'static str;
}

A trait applied to all Kubernetes resources that can be part of a corresponding list.

Required Associated Constants

const LIST_KIND: &'static str

The kind of the list type of the resource.

This is the string used in the kind field of the list type’s serialized form.

Object Safety

This trait is not object safe.