Module k8s_openapi::api::admissionregistration::v1

Structs

MutatingWebhook

MutatingWebhook describes an admission webhook and the resources and operations it applies to.

MutatingWebhookConfiguration

MutatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and may change the object.

RuleWithOperations

RuleWithOperations is a tuple of Operations and Resources. It is recommended to make sure that all the tuple expansions are valid.

ServiceReference

ServiceReference holds a reference to Service.legacy.k8s.io

ValidatingWebhook

ValidatingWebhook describes an admission webhook and the resources and operations it applies to.

ValidatingWebhookConfiguration

ValidatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and object without changing it.

WebhookClientConfig

WebhookClientConfig contains the information to make a TLS connection with the webhook

Enums

ReadMutatingWebhookConfigurationResponse

Use <ReadMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::read

ReadValidatingWebhookConfigurationResponse

Use <ReadValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::read