k8s_openapi::api::admissionregistration
Module v1alpha1

Structs

ApplyConfiguration

ApplyConfiguration defines the desired configuration values of an object.

JSONPatch

JSONPatch defines a JSON Patch.

MatchCondition

MatchResources

MatchResources decides whether to run the admission control policy on an object based on whether it meets the match criteria. The exclude rules take precedence over include rules (if a resource matches both, it is excluded)

MutatingAdmissionPolicy

MutatingAdmissionPolicy describes the definition of an admission mutation policy that mutates the object coming into admission chain.

MutatingAdmissionPolicyBinding

MutatingAdmissionPolicyBinding binds the MutatingAdmissionPolicy with parametrized resources. MutatingAdmissionPolicyBinding and the optional parameter resource together define how cluster administrators configure policies for clusters.

MutatingAdmissionPolicyBindingSpec

MutatingAdmissionPolicyBindingSpec is the specification of the MutatingAdmissionPolicyBinding.

MutatingAdmissionPolicySpec

MutatingAdmissionPolicySpec is the specification of the desired behavior of the admission policy.

Mutation

Mutation specifies the CEL expression which is used to apply the Mutation.

NamedRuleWithOperations

NamedRuleWithOperations is a tuple of Operations and Resources with ResourceNames.

ParamKind

ParamKind is a tuple of Group Kind and Version.

ParamRef

ParamRef describes how to locate the params to be used as input to expressions of rules applied by a policy binding.

Variable

Variable is the definition of a variable that is used for composition.