Module k8s_openapi::api::admissionregistration::v1alpha1

Structs

MatchResources

MatchResources decides whether to run the admission control policy on an object based on whether it meets the match criteria. The exclude rules take precedence over include rules (if a resource matches both, it is excluded)

NamedRuleWithOperations

NamedRuleWithOperations is a tuple of Operations and Resources with ResourceNames.

ParamKind

ParamKind is a tuple of Group Kind and Version.

ParamRef

ParamRef references a parameter resource

ValidatingAdmissionPolicy

ValidatingAdmissionPolicy describes the definition of an admission validation policy that accepts or rejects an object without changing it.

ValidatingAdmissionPolicyBinding

ValidatingAdmissionPolicyBinding binds the ValidatingAdmissionPolicy with paramerized resources. ValidatingAdmissionPolicyBinding and parameter CRDs together define how cluster administrators configure policies for clusters.

ValidatingAdmissionPolicyBindingSpec

ValidatingAdmissionPolicyBindingSpec is the specification of the ValidatingAdmissionPolicyBinding.

ValidatingAdmissionPolicySpec

ValidatingAdmissionPolicySpec is the specification of the desired behavior of the AdmissionPolicy.

Validation

Validation specifies the CEL expression which is used to apply the validation.

Enums

ReadValidatingAdmissionPolicyBindingResponse

Use <ReadValidatingAdmissionPolicyBindingResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingAdmissionPolicyBinding::read

ReadValidatingAdmissionPolicyResponse

Use <ReadValidatingAdmissionPolicyResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingAdmissionPolicy::read