Module k8s_openapi::api::admissionregistration::v1beta1

Structs

MutatingWebhook
MutatingWebhook describes an admission webhook and the resources and operations it applies to.

MutatingWebhookConfiguration
MutatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and may change the object. Deprecated in v1.16, planned for removal in v1.19. Use admissionregistration.k8s.io/v1 MutatingWebhookConfiguration instead.

ReadMutatingWebhookConfigurationOptional
Optional parameters of MutatingWebhookConfiguration::read_mutating_webhook_configuration

ReadValidatingWebhookConfigurationOptional
Optional parameters of ValidatingWebhookConfiguration::read_validating_webhook_configuration

RuleWithOperations
RuleWithOperations is a tuple of Operations and Resources. It is recommended to make sure that all the tuple expansions are valid.

ServiceReference
ServiceReference holds a reference to Service.legacy.k8s.io

ValidatingWebhook
ValidatingWebhook describes an admission webhook and the resources and operations it applies to.

ValidatingWebhookConfiguration
ValidatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and object without changing it. Deprecated in v1.16, planned for removal in v1.19. Use admissionregistration.k8s.io/v1 ValidatingWebhookConfiguration instead.

WebhookClientConfig
WebhookClientConfig contains the information to make a TLS connection with the webhook

Enums

ReadMutatingWebhookConfigurationResponse
Use <ReadMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::read_mutating_webhook_configuration

ReadValidatingWebhookConfigurationResponse
Use <ReadValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::read_validating_webhook_configuration