Module k8s_openapi::api::admissionregistration::v1beta1

Structs

CreateMutatingWebhookConfigurationOptional
Optional parameters of MutatingWebhookConfiguration::create_mutating_webhook_configuration

CreateValidatingWebhookConfigurationOptional
Optional parameters of ValidatingWebhookConfiguration::create_validating_webhook_configuration

MutatingWebhook
MutatingWebhook describes an admission webhook and the resources and operations it applies to.

MutatingWebhookConfiguration
MutatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and may change the object. Deprecated in v1.16, planned for removal in v1.19. Use admissionregistration.k8s.io/v1 MutatingWebhookConfiguration instead.

MutatingWebhookConfigurationList
MutatingWebhookConfigurationList is a list of MutatingWebhookConfiguration.

ReadMutatingWebhookConfigurationOptional
Optional parameters of MutatingWebhookConfiguration::read_mutating_webhook_configuration

ReadValidatingWebhookConfigurationOptional
Optional parameters of ValidatingWebhookConfiguration::read_validating_webhook_configuration

ReplaceMutatingWebhookConfigurationOptional
Optional parameters of MutatingWebhookConfiguration::replace_mutating_webhook_configuration

ReplaceValidatingWebhookConfigurationOptional
Optional parameters of ValidatingWebhookConfiguration::replace_validating_webhook_configuration

RuleWithOperations
RuleWithOperations is a tuple of Operations and Resources. It is recommended to make sure that all the tuple expansions are valid.

ServiceReference
ServiceReference holds a reference to Service.legacy.k8s.io

ValidatingWebhook
ValidatingWebhook describes an admission webhook and the resources and operations it applies to.

ValidatingWebhookConfiguration
ValidatingWebhookConfiguration describes the configuration of and admission webhook that accept or reject and object without changing it. Deprecated in v1.16, planned for removal in v1.19. Use admissionregistration.k8s.io/v1 ValidatingWebhookConfiguration instead.

ValidatingWebhookConfigurationList
ValidatingWebhookConfigurationList is a list of ValidatingWebhookConfiguration.

WebhookClientConfig
WebhookClientConfig contains the information to make a TLS connection with the webhook

Enums

CreateMutatingWebhookConfigurationResponse
Use <CreateMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::create_mutating_webhook_configuration

CreateValidatingWebhookConfigurationResponse
Use <CreateValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::create_validating_webhook_configuration

DeleteCollectionMutatingWebhookConfigurationResponse
Use <DeleteCollectionMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::delete_collection_mutating_webhook_configuration

DeleteCollectionValidatingWebhookConfigurationResponse
Use <DeleteCollectionValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::delete_collection_validating_webhook_configuration

DeleteMutatingWebhookConfigurationResponse
Use <DeleteMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::delete_mutating_webhook_configuration

DeleteValidatingWebhookConfigurationResponse
Use <DeleteValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::delete_validating_webhook_configuration

ListMutatingWebhookConfigurationResponse
Use <ListMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::list_mutating_webhook_configuration

ListValidatingWebhookConfigurationResponse
Use <ListValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::list_validating_webhook_configuration

PatchMutatingWebhookConfigurationResponse
Use <PatchMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::patch_mutating_webhook_configuration

PatchValidatingWebhookConfigurationResponse
Use <PatchValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::patch_validating_webhook_configuration

ReadMutatingWebhookConfigurationResponse
Use <ReadMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::read_mutating_webhook_configuration

ReadValidatingWebhookConfigurationResponse
Use <ReadValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::read_validating_webhook_configuration

ReplaceMutatingWebhookConfigurationResponse
Use <ReplaceMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::replace_mutating_webhook_configuration

ReplaceValidatingWebhookConfigurationResponse
Use <ReplaceValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::replace_validating_webhook_configuration

WatchMutatingWebhookConfigurationResponse
Use <WatchMutatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of MutatingWebhookConfiguration::watch_mutating_webhook_configuration

WatchValidatingWebhookConfigurationResponse
Use <WatchValidatingWebhookConfigurationResponse as Response>::try_from_parts to parse the HTTP response body of ValidatingWebhookConfiguration::watch_validating_webhook_configuration