Module k8s_openapi::api::apiserverinternal::v1alpha1

Structs

ServerStorageVersion

An API server instance reports the version it can decode and the version it encodes objects to when persisting objects in the backend.

StorageVersion

Storage version of a specific resource.

StorageVersionCondition

Describes the state of the storageVersion at a certain point.

StorageVersionSpec

StorageVersionSpec is an empty spec.

StorageVersionStatus

API server instances report the versions they can decode and the version they encode objects to when persisting objects in the backend.

Enums

ReadStorageVersionResponse

Use <ReadStorageVersionResponse as Response>::try_from_parts to parse the HTTP response body of StorageVersion::read

ReadStorageVersionStatusResponse

Use <ReadStorageVersionStatusResponse as Response>::try_from_parts to parse the HTTP response body of StorageVersion::read_status