Module k8s_openapi::api::apps::v1

Structs

ControllerRevision
ControllerRevision implements an immutable snapshot of state data. Clients are responsible for serializing and deserializing the objects that contain their internal state. Once a ControllerRevision has been successfully created, it can not be updated. The API Server will fail validation of all requests that attempt to mutate the Data field. ControllerRevisions may, however, be deleted. Note that, due to its use by both the DaemonSet and StatefulSet controllers for update and rollback, this object is beta. However, it may be subject to name and representation changes in future releases, and clients should not depend on its stability. It is primarily for internal use by controllers.

DaemonSet
DaemonSet represents the configuration of a daemon set.

DaemonSetCondition
DaemonSetCondition describes the state of a DaemonSet at a certain point.

DaemonSetSpec
DaemonSetSpec is the specification of a daemon set.

DaemonSetStatus
DaemonSetStatus represents the current status of a daemon set.

DaemonSetUpdateStrategy
DaemonSetUpdateStrategy is a struct used to control the update strategy for a DaemonSet.

Deployment
Deployment enables declarative updates for Pods and ReplicaSets.

DeploymentCondition
DeploymentCondition describes the state of a deployment at a certain point.

DeploymentSpec
DeploymentSpec is the specification of the desired behavior of the Deployment.

DeploymentStatus
DeploymentStatus is the most recently observed status of the Deployment.

DeploymentStrategy
DeploymentStrategy describes how to replace existing pods with new ones.

ReadNamespacedControllerRevisionOptional
Optional parameters of ControllerRevision::read_namespaced_controller_revision

ReadNamespacedDaemonSetOptional
Optional parameters of DaemonSet::read_namespaced_daemon_set

ReadNamespacedDaemonSetStatusOptional
Optional parameters of DaemonSet::read_namespaced_daemon_set_status

ReadNamespacedDeploymentOptional
Optional parameters of Deployment::read_namespaced_deployment

ReadNamespacedDeploymentStatusOptional
Optional parameters of Deployment::read_namespaced_deployment_status

ReadNamespacedReplicaSetOptional
Optional parameters of ReplicaSet::read_namespaced_replica_set

ReadNamespacedReplicaSetStatusOptional
Optional parameters of ReplicaSet::read_namespaced_replica_set_status

ReadNamespacedStatefulSetOptional
Optional parameters of StatefulSet::read_namespaced_stateful_set

ReadNamespacedStatefulSetStatusOptional
Optional parameters of StatefulSet::read_namespaced_stateful_set_status

ReplicaSet
ReplicaSet ensures that a specified number of pod replicas are running at any given time.

ReplicaSetCondition
ReplicaSetCondition describes the state of a replica set at a certain point.

ReplicaSetSpec
ReplicaSetSpec is the specification of a ReplicaSet.

ReplicaSetStatus
ReplicaSetStatus represents the current status of a ReplicaSet.

RollingUpdateDaemonSet
Spec to control the desired behavior of daemon set rolling update.

RollingUpdateDeployment
Spec to control the desired behavior of rolling update.

RollingUpdateStatefulSetStrategy
RollingUpdateStatefulSetStrategy is used to communicate parameter for RollingUpdateStatefulSetStrategyType.

StatefulSet
StatefulSet represents a set of pods with consistent identities. Identities are defined as:

StatefulSetCondition
StatefulSetCondition describes the state of a statefulset at a certain point.

StatefulSetSpec
A StatefulSetSpec is the specification of a StatefulSet.

StatefulSetStatus
StatefulSetStatus represents the current state of a StatefulSet.

StatefulSetUpdateStrategy
StatefulSetUpdateStrategy indicates the strategy that the StatefulSet controller will use to perform updates. It includes any additional parameters necessary to perform the update for the indicated strategy.

Enums

ReadNamespacedControllerRevisionResponse
Use <ReadNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::read_namespaced_controller_revision

ReadNamespacedDaemonSetResponse
Use <ReadNamespacedDaemonSetResponse as Response>::try_from_parts to parse the HTTP response body of DaemonSet::read_namespaced_daemon_set

ReadNamespacedDaemonSetStatusResponse
Use <ReadNamespacedDaemonSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of DaemonSet::read_namespaced_daemon_set_status

ReadNamespacedDeploymentResponse
Use <ReadNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::read_namespaced_deployment

ReadNamespacedDeploymentStatusResponse
Use <ReadNamespacedDeploymentStatusResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::read_namespaced_deployment_status

ReadNamespacedReplicaSetResponse
Use <ReadNamespacedReplicaSetResponse as Response>::try_from_parts to parse the HTTP response body of ReplicaSet::read_namespaced_replica_set

ReadNamespacedReplicaSetStatusResponse
Use <ReadNamespacedReplicaSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of ReplicaSet::read_namespaced_replica_set_status

ReadNamespacedStatefulSetResponse
Use <ReadNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::read_namespaced_stateful_set

ReadNamespacedStatefulSetStatusResponse
Use <ReadNamespacedStatefulSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::read_namespaced_stateful_set_status