Module k8s_openapi::api::apps::v1beta1

Structs

ControllerRevision
DEPRECATED - This group version of ControllerRevision is deprecated by apps/v1beta2/ControllerRevision. See the release notes for more information. ControllerRevision implements an immutable snapshot of state data. Clients are responsible for serializing and deserializing the objects that contain their internal state. Once a ControllerRevision has been successfully created, it can not be updated. The API Server will fail validation of all requests that attempt to mutate the Data field. ControllerRevisions may, however, be deleted. Note that, due to its use by both the DaemonSet and StatefulSet controllers for update and rollback, this object is beta. However, it may be subject to name and representation changes in future releases, and clients should not depend on its stability. It is primarily for internal use by controllers.

ControllerRevisionList
ControllerRevisionList is a resource containing a list of ControllerRevision objects.

CreateNamespacedControllerRevisionOptional
Optional parameters of ControllerRevision::create_namespaced_controller_revision

CreateNamespacedDeploymentOptional
Optional parameters of Deployment::create_namespaced_deployment

CreateNamespacedDeploymentRollbackOptional
Optional parameters of DeploymentRollback::create_namespaced_deployment_rollback

CreateNamespacedStatefulSetOptional
Optional parameters of StatefulSet::create_namespaced_stateful_set

Deployment
DEPRECATED - This group version of Deployment is deprecated by apps/v1beta2/Deployment. See the release notes for more information. Deployment enables declarative updates for Pods and ReplicaSets.

DeploymentCondition
DeploymentCondition describes the state of a deployment at a certain point.

DeploymentList
DeploymentList is a list of Deployments.

DeploymentRollback
DEPRECATED. DeploymentRollback stores the information required to rollback a deployment.

DeploymentSpec
DeploymentSpec is the specification of the desired behavior of the Deployment.

DeploymentStatus
DeploymentStatus is the most recently observed status of the Deployment.

DeploymentStrategy
DeploymentStrategy describes how to replace existing pods with new ones.

ReadNamespacedControllerRevisionOptional
Optional parameters of ControllerRevision::read_namespaced_controller_revision

ReadNamespacedDeploymentOptional
Optional parameters of Deployment::read_namespaced_deployment

ReadNamespacedDeploymentScaleOptional
Optional parameters of Scale::read_namespaced_deployment_scale

ReadNamespacedDeploymentStatusOptional
Optional parameters of Deployment::read_namespaced_deployment_status

ReadNamespacedStatefulSetOptional
Optional parameters of StatefulSet::read_namespaced_stateful_set

ReadNamespacedStatefulSetScaleOptional
Optional parameters of Scale::read_namespaced_stateful_set_scale

ReadNamespacedStatefulSetStatusOptional
Optional parameters of StatefulSet::read_namespaced_stateful_set_status

ReplaceNamespacedControllerRevisionOptional
Optional parameters of ControllerRevision::replace_namespaced_controller_revision

ReplaceNamespacedDeploymentOptional
Optional parameters of Deployment::replace_namespaced_deployment

ReplaceNamespacedDeploymentScaleOptional
Optional parameters of Scale::replace_namespaced_deployment_scale

ReplaceNamespacedDeploymentStatusOptional
Optional parameters of Deployment::replace_namespaced_deployment_status

ReplaceNamespacedStatefulSetOptional
Optional parameters of StatefulSet::replace_namespaced_stateful_set

ReplaceNamespacedStatefulSetScaleOptional
Optional parameters of Scale::replace_namespaced_stateful_set_scale

ReplaceNamespacedStatefulSetStatusOptional
Optional parameters of StatefulSet::replace_namespaced_stateful_set_status

RollbackConfig
DEPRECATED.

RollingUpdateDeployment
Spec to control the desired behavior of rolling update.

RollingUpdateStatefulSetStrategy
RollingUpdateStatefulSetStrategy is used to communicate parameter for RollingUpdateStatefulSetStrategyType.

Scale
Scale represents a scaling request for a resource.

ScaleSpec
ScaleSpec describes the attributes of a scale subresource

ScaleStatus
ScaleStatus represents the current status of a scale subresource.

StatefulSet
DEPRECATED - This group version of StatefulSet is deprecated by apps/v1beta2/StatefulSet. See the release notes for more information. StatefulSet represents a set of pods with consistent identities. Identities are defined as:

StatefulSetCondition
StatefulSetCondition describes the state of a statefulset at a certain point.

StatefulSetList
StatefulSetList is a collection of StatefulSets.

StatefulSetSpec
A StatefulSetSpec is the specification of a StatefulSet.

StatefulSetStatus
StatefulSetStatus represents the current state of a StatefulSet.

StatefulSetUpdateStrategy
StatefulSetUpdateStrategy indicates the strategy that the StatefulSet controller will use to perform updates. It includes any additional parameters necessary to perform the update for the indicated strategy.

Enums

CreateNamespacedControllerRevisionResponse
Use <CreateNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::create_namespaced_controller_revision

CreateNamespacedDeploymentResponse
Use <CreateNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::create_namespaced_deployment

CreateNamespacedDeploymentRollbackResponse
Use <CreateNamespacedDeploymentRollbackResponse as Response>::try_from_parts to parse the HTTP response body of DeploymentRollback::create_namespaced_deployment_rollback

CreateNamespacedStatefulSetResponse
Use <CreateNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::create_namespaced_stateful_set

DeleteCollectionNamespacedControllerRevisionResponse
Use <DeleteCollectionNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::delete_collection_namespaced_controller_revision

DeleteCollectionNamespacedDeploymentResponse
Use <DeleteCollectionNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::delete_collection_namespaced_deployment

DeleteCollectionNamespacedStatefulSetResponse
Use <DeleteCollectionNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::delete_collection_namespaced_stateful_set

DeleteNamespacedControllerRevisionResponse
Use <DeleteNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::delete_namespaced_controller_revision

DeleteNamespacedDeploymentResponse
Use <DeleteNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::delete_namespaced_deployment

DeleteNamespacedStatefulSetResponse
Use <DeleteNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::delete_namespaced_stateful_set

ListControllerRevisionForAllNamespacesResponse
Use <ListControllerRevisionForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::list_controller_revision_for_all_namespaces

ListDeploymentForAllNamespacesResponse
Use <ListDeploymentForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::list_deployment_for_all_namespaces

ListNamespacedControllerRevisionResponse
Use <ListNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::list_namespaced_controller_revision

ListNamespacedDeploymentResponse
Use <ListNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::list_namespaced_deployment

ListNamespacedStatefulSetResponse
Use <ListNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::list_namespaced_stateful_set

ListStatefulSetForAllNamespacesResponse
Use <ListStatefulSetForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::list_stateful_set_for_all_namespaces

PatchNamespacedControllerRevisionResponse
Use <PatchNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::patch_namespaced_controller_revision

PatchNamespacedDeploymentResponse
Use <PatchNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::patch_namespaced_deployment

PatchNamespacedDeploymentScaleResponse
Use <PatchNamespacedDeploymentScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::patch_namespaced_deployment_scale

PatchNamespacedDeploymentStatusResponse
Use <PatchNamespacedDeploymentStatusResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::patch_namespaced_deployment_status

PatchNamespacedStatefulSetResponse
Use <PatchNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::patch_namespaced_stateful_set

PatchNamespacedStatefulSetScaleResponse
Use <PatchNamespacedStatefulSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::patch_namespaced_stateful_set_scale

PatchNamespacedStatefulSetStatusResponse
Use <PatchNamespacedStatefulSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::patch_namespaced_stateful_set_status

ReadNamespacedControllerRevisionResponse
Use <ReadNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::read_namespaced_controller_revision

ReadNamespacedDeploymentResponse
Use <ReadNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::read_namespaced_deployment

ReadNamespacedDeploymentScaleResponse
Use <ReadNamespacedDeploymentScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_deployment_scale

ReadNamespacedDeploymentStatusResponse
Use <ReadNamespacedDeploymentStatusResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::read_namespaced_deployment_status

ReadNamespacedStatefulSetResponse
Use <ReadNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::read_namespaced_stateful_set

ReadNamespacedStatefulSetScaleResponse
Use <ReadNamespacedStatefulSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_stateful_set_scale

ReadNamespacedStatefulSetStatusResponse
Use <ReadNamespacedStatefulSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::read_namespaced_stateful_set_status

ReplaceNamespacedControllerRevisionResponse
Use <ReplaceNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::replace_namespaced_controller_revision

ReplaceNamespacedDeploymentResponse
Use <ReplaceNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::replace_namespaced_deployment

ReplaceNamespacedDeploymentScaleResponse
Use <ReplaceNamespacedDeploymentScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::replace_namespaced_deployment_scale

ReplaceNamespacedDeploymentStatusResponse
Use <ReplaceNamespacedDeploymentStatusResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::replace_namespaced_deployment_status

ReplaceNamespacedStatefulSetResponse
Use <ReplaceNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::replace_namespaced_stateful_set

ReplaceNamespacedStatefulSetScaleResponse
Use <ReplaceNamespacedStatefulSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::replace_namespaced_stateful_set_scale

ReplaceNamespacedStatefulSetStatusResponse
Use <ReplaceNamespacedStatefulSetStatusResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::replace_namespaced_stateful_set_status

WatchControllerRevisionForAllNamespacesResponse
Use <WatchControllerRevisionForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::watch_controller_revision_for_all_namespaces

WatchDeploymentForAllNamespacesResponse
Use <WatchDeploymentForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::watch_deployment_for_all_namespaces

WatchNamespacedControllerRevisionResponse
Use <WatchNamespacedControllerRevisionResponse as Response>::try_from_parts to parse the HTTP response body of ControllerRevision::watch_namespaced_controller_revision

WatchNamespacedDeploymentResponse
Use <WatchNamespacedDeploymentResponse as Response>::try_from_parts to parse the HTTP response body of Deployment::watch_namespaced_deployment

WatchNamespacedStatefulSetResponse
Use <WatchNamespacedStatefulSetResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::watch_namespaced_stateful_set

WatchStatefulSetForAllNamespacesResponse
Use <WatchStatefulSetForAllNamespacesResponse as Response>::try_from_parts to parse the HTTP response body of StatefulSet::watch_stateful_set_for_all_namespaces