Module k8s_openapi::api::auditregistration::v1alpha1

Structs

AuditSink
AuditSink represents a cluster level audit sink

AuditSinkList
AuditSinkList is a list of AuditSink items.

AuditSinkSpec
AuditSinkSpec holds the spec for the audit sink

CreateAuditSinkOptional
Optional parameters of AuditSink::create_audit_sink

Policy
Policy defines the configuration of how audit events are logged

ReadAuditSinkOptional
Optional parameters of AuditSink::read_audit_sink

ReplaceAuditSinkOptional
Optional parameters of AuditSink::replace_audit_sink

ServiceReference
ServiceReference holds a reference to Service.legacy.k8s.io

Webhook
Webhook holds the configuration of the webhook

WebhookClientConfig
WebhookClientConfig contains the information to make a connection with the webhook

WebhookThrottleConfig
WebhookThrottleConfig holds the configuration for throttling events

Enums

CreateAuditSinkResponse
Use <CreateAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::create_audit_sink

DeleteAuditSinkResponse
Use <DeleteAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::delete_audit_sink

DeleteCollectionAuditSinkResponse
Use <DeleteCollectionAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::delete_collection_audit_sink

ListAuditSinkResponse
Use <ListAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::list_audit_sink

PatchAuditSinkResponse
Use <PatchAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::patch_audit_sink

ReadAuditSinkResponse
Use <ReadAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::read_audit_sink

ReplaceAuditSinkResponse
Use <ReplaceAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::replace_audit_sink

WatchAuditSinkResponse
Use <WatchAuditSinkResponse as Response>::try_from_parts to parse the HTTP response body of AuditSink::watch_audit_sink