Module k8s_openapi::api::authentication::v1

Structs

BoundObjectReference
BoundObjectReference is a reference to an object that a token is bound to.

CreateNamespacedServiceAccountTokenOptional
Optional parameters of TokenRequest::create_namespaced_service_account_token

CreateTokenReviewOptional
Optional parameters of TokenReview::create_token_review

TokenRequest
TokenRequest requests a token for a given service account.

TokenRequestSpec
TokenRequestSpec contains client provided parameters of a token request.

TokenRequestStatus
TokenRequestStatus is the result of a token request.

TokenReview
TokenReview attempts to authenticate a token to a known user. Note: TokenReview requests may be cached by the webhook token authenticator plugin in the kube-apiserver.

TokenReviewSpec
TokenReviewSpec is a description of the token authentication request.

TokenReviewStatus
TokenReviewStatus is the result of the token authentication request.

UserInfo
UserInfo holds the information about the user needed to implement the user.Info interface.

Enums

CreateNamespacedServiceAccountTokenResponse
Use <CreateNamespacedServiceAccountTokenResponse as Response>::try_from_parts to parse the HTTP response body of TokenRequest::create_namespaced_service_account_token

CreateTokenReviewResponse
Use <CreateTokenReviewResponse as Response>::try_from_parts to parse the HTTP response body of TokenReview::create_token_review