Module k8s_openapi::api::autoscaling::v1

Structs

CrossVersionObjectReference

CrossVersionObjectReference contains enough information to let you identify the referred resource.

HorizontalPodAutoscaler

configuration of a horizontal pod autoscaler.

HorizontalPodAutoscalerSpec

specification of a horizontal pod autoscaler.

HorizontalPodAutoscalerStatus

current status of a horizontal pod autoscaler

ReadNamespacedDeploymentScaleOptional

Optional parameters of Scale::read_namespaced_deployment_scale

ReadNamespacedHorizontalPodAutoscalerOptional

Optional parameters of HorizontalPodAutoscaler::read_namespaced_horizontal_pod_autoscaler

ReadNamespacedHorizontalPodAutoscalerStatusOptional

Optional parameters of HorizontalPodAutoscaler::read_namespaced_horizontal_pod_autoscaler_status

ReadNamespacedReplicaSetScaleOptional

Optional parameters of Scale::read_namespaced_replica_set_scale

ReadNamespacedReplicationControllerScaleOptional

Optional parameters of Scale::read_namespaced_replication_controller_scale

ReadNamespacedStatefulSetScaleOptional

Optional parameters of Scale::read_namespaced_stateful_set_scale

Scale

Scale represents a scaling request for a resource.

ScaleSpec

ScaleSpec describes the attributes of a scale subresource.

ScaleStatus

ScaleStatus represents the current status of a scale subresource.

Enums

ReadNamespacedDeploymentScaleResponse

Use <ReadNamespacedDeploymentScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_deployment_scale

ReadNamespacedHorizontalPodAutoscalerResponse

Use <ReadNamespacedHorizontalPodAutoscalerResponse as Response>::try_from_parts to parse the HTTP response body of HorizontalPodAutoscaler::read_namespaced_horizontal_pod_autoscaler

ReadNamespacedHorizontalPodAutoscalerStatusResponse

Use <ReadNamespacedHorizontalPodAutoscalerStatusResponse as Response>::try_from_parts to parse the HTTP response body of HorizontalPodAutoscaler::read_namespaced_horizontal_pod_autoscaler_status

ReadNamespacedReplicaSetScaleResponse

Use <ReadNamespacedReplicaSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_replica_set_scale

ReadNamespacedReplicationControllerScaleResponse

Use <ReadNamespacedReplicationControllerScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_replication_controller_scale

ReadNamespacedStatefulSetScaleResponse

Use <ReadNamespacedStatefulSetScaleResponse as Response>::try_from_parts to parse the HTTP response body of Scale::read_namespaced_stateful_set_scale