Module k8s_openapi::api::autoscaling::v1

Structs

CrossVersionObjectReference

CrossVersionObjectReference contains enough information to let you identify the referred resource.

HorizontalPodAutoscaler

configuration of a horizontal pod autoscaler.

HorizontalPodAutoscalerSpec

specification of a horizontal pod autoscaler.

HorizontalPodAutoscalerStatus

current status of a horizontal pod autoscaler

Scale

Scale represents a scaling request for a resource.

ScaleSpec

ScaleSpec describes the attributes of a scale subresource.

ScaleStatus

ScaleStatus represents the current status of a scale subresource.