k8s_openapi::api::batch

Module v1

Structs

CronJob

CronJob represents the configuration of a single cron job.

CronJobSpec

CronJobSpec describes how the job execution will look like and when it will actually run.

CronJobStatus

CronJobStatus represents the current state of a cron job.

Job

Job represents the configuration of a single job.

JobCondition

JobCondition describes current state of a job.

JobSpec

JobSpec describes how the job execution will look like.

JobStatus

JobStatus represents the current state of a Job.

JobTemplateSpec

JobTemplateSpec describes the data a Job should have when created from a template

PodFailurePolicy

PodFailurePolicy describes how failed pods influence the backoffLimit.

PodFailurePolicyOnExitCodesRequirement

PodFailurePolicyOnExitCodesRequirement describes the requirement for handling a failed pod based on its container exit codes. In particular, it lookups the .state.terminated.exitCode for each app container and init container status, represented by the .status.containerStatuses and .status.initContainerStatuses fields in the Pod status, respectively. Containers completed with success (exit code 0) are excluded from the requirement check.

PodFailurePolicyOnPodConditionsPattern

PodFailurePolicyOnPodConditionsPattern describes a pattern for matching an actual pod condition type.

PodFailurePolicyRule

PodFailurePolicyRule describes how a pod failure is handled when the requirements are met. One of onExitCodes and onPodConditions, but not both, can be used in each rule.

SuccessPolicy

SuccessPolicy describes when a Job can be declared as succeeded based on the success of some indexes.

SuccessPolicyRule

SuccessPolicyRule describes rule for declaring a Job as succeeded. Each rule must have at least one of the “succeededIndexes” or “succeededCount” specified.

UncountedTerminatedPods

UncountedTerminatedPods holds UIDs of Pods that have terminated but haven’t been accounted in Job status counters.