Module k8s_openapi::api::batch::v1

Structs

CronJob
CronJob represents the configuration of a single cron job.

CronJobSpec
CronJobSpec describes how the job execution will look like and when it will actually run.

CronJobStatus
CronJobStatus represents the current state of a cron job.

Job
Job represents the configuration of a single job.

JobCondition
JobCondition describes current state of a job.

JobSpec
JobSpec describes how the job execution will look like.

JobStatus
JobStatus represents the current state of a Job.

JobTemplateSpec
JobTemplateSpec describes the data a Job should have when created from a template

ReadNamespacedCronJobOptional
Optional parameters of CronJob::read_namespaced_cron_job

ReadNamespacedCronJobStatusOptional
Optional parameters of CronJob::read_namespaced_cron_job_status

ReadNamespacedJobOptional
Optional parameters of Job::read_namespaced_job

ReadNamespacedJobStatusOptional
Optional parameters of Job::read_namespaced_job_status

Enums

ReadNamespacedCronJobResponse
Use <ReadNamespacedCronJobResponse as Response>::try_from_parts to parse the HTTP response body of CronJob::read_namespaced_cron_job

ReadNamespacedCronJobStatusResponse
Use <ReadNamespacedCronJobStatusResponse as Response>::try_from_parts to parse the HTTP response body of CronJob::read_namespaced_cron_job_status

ReadNamespacedJobResponse
Use <ReadNamespacedJobResponse as Response>::try_from_parts to parse the HTTP response body of Job::read_namespaced_job

ReadNamespacedJobStatusResponse
Use <ReadNamespacedJobStatusResponse as Response>::try_from_parts to parse the HTTP response body of Job::read_namespaced_job_status