Module k8s_openapi::api::certificates::v1beta1

Structs

CertificateSigningRequest
Describes a certificate signing request

CertificateSigningRequestCondition

CertificateSigningRequestList

CertificateSigningRequestSpec
This information is immutable after the request is created. Only the Request and Usages fields can be set on creation, other fields are derived by Kubernetes and cannot be modified by users.

CertificateSigningRequestStatus

CreateCertificateSigningRequestOptional
Optional parameters of CertificateSigningRequest::create_certificate_signing_request

ReadCertificateSigningRequestOptional
Optional parameters of CertificateSigningRequest::read_certificate_signing_request

ReadCertificateSigningRequestStatusOptional
Optional parameters of CertificateSigningRequest::read_certificate_signing_request_status

ReplaceCertificateSigningRequestApprovalOptional
Optional parameters of CertificateSigningRequest::replace_certificate_signing_request_approval

ReplaceCertificateSigningRequestOptional
Optional parameters of CertificateSigningRequest::replace_certificate_signing_request

ReplaceCertificateSigningRequestStatusOptional
Optional parameters of CertificateSigningRequest::replace_certificate_signing_request_status

Enums

CreateCertificateSigningRequestResponse
Use <CreateCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::create_certificate_signing_request

DeleteCertificateSigningRequestResponse
Use <DeleteCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::delete_certificate_signing_request

DeleteCollectionCertificateSigningRequestResponse
Use <DeleteCollectionCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::delete_collection_certificate_signing_request

ListCertificateSigningRequestResponse
Use <ListCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::list_certificate_signing_request

PatchCertificateSigningRequestResponse
Use <PatchCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::patch_certificate_signing_request

PatchCertificateSigningRequestStatusResponse
Use <PatchCertificateSigningRequestStatusResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::patch_certificate_signing_request_status

ReadCertificateSigningRequestResponse
Use <ReadCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::read_certificate_signing_request

ReadCertificateSigningRequestStatusResponse
Use <ReadCertificateSigningRequestStatusResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::read_certificate_signing_request_status

ReplaceCertificateSigningRequestApprovalResponse
Use <ReplaceCertificateSigningRequestApprovalResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::replace_certificate_signing_request_approval

ReplaceCertificateSigningRequestResponse
Use <ReplaceCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::replace_certificate_signing_request

ReplaceCertificateSigningRequestStatusResponse
Use <ReplaceCertificateSigningRequestStatusResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::replace_certificate_signing_request_status

WatchCertificateSigningRequestResponse
Use <WatchCertificateSigningRequestResponse as Response>::try_from_parts to parse the HTTP response body of CertificateSigningRequest::watch_certificate_signing_request