Struct k8s_openapi::api::core::v1::EphemeralContainer

pub struct EphemeralContainer {

    pub args: Option<Vec<String>>,
    pub command: Option<Vec<String>>,
    pub env: Option<Vec<EnvVar>>,
    pub env_from: Option<Vec<EnvFromSource>>,
    pub image: Option<String>,
    pub image_pull_policy: Option<String>,
    pub lifecycle: Option<Lifecycle>,
    pub liveness_probe: Option<Probe>,
    pub name: String,
    pub ports: Option<Vec<ContainerPort>>,
    pub readiness_probe: Option<Probe>,
    pub resources: Option<ResourceRequirements>,
    pub security_context: Option<SecurityContext>,
    pub startup_probe: Option<Probe>,
    pub stdin: Option<bool>,
    pub stdin_once: Option<bool>,
    pub target_container_name: Option<String>,
    pub termination_message_path: Option<String>,
    pub termination_message_policy: Option<String>,
    pub tty: Option<bool>,
    pub volume_devices: Option<Vec<VolumeDevice>>,
    pub volume_mounts: Option<Vec<VolumeMount>>,
    pub working_dir: Option<String>,

}

An EphemeralContainer is a temporary container that you may add to an existing Pod for user-initiated activities such as debugging. Ephemeral containers have no resource or scheduling guarantees, and they will not be restarted when they exit or when a Pod is removed or restarted. The kubelet may evict a Pod if an ephemeral container causes the Pod to exceed its resource allocation.

To add an ephemeral container, use the ephemeralcontainers subresource of an existing Pod. Ephemeral containers may not be removed or restarted.

Fields
args: Option<Vec<String>>

Arguments to the entrypoint. The image’s CMD is used if this is not provided. Variable references $(VAR_NAME) are expanded using the container’s environment. If a variable cannot be resolved, the reference in the input string will be unchanged. Double $$ are reduced to a single $, which allows for escaping the $(VAR_NAME) syntax: i.e. “$$(VAR_NAME)” will produce the string literal “$(VAR_NAME)”. Escaped references will never be expanded, regardless of whether the variable exists or not. Cannot be updated. More info: https://kubernetes.io/docs/tasks/inject-data-application/define-command-argument-container/#running-a-command-in-a-shell

command: Option<Vec<String>>

Entrypoint array. Not executed within a shell. The image’s ENTRYPOINT is used if this is not provided. Variable references $(VAR_NAME) are expanded using the container’s environment. If a variable cannot be resolved, the reference in the input string will be unchanged. Double $$ are reduced to a single $, which allows for escaping the $(VAR_NAME) syntax: i.e. “$$(VAR_NAME)” will produce the string literal “$(VAR_NAME)”. Escaped references will never be expanded, regardless of whether the variable exists or not. Cannot be updated. More info: https://kubernetes.io/docs/tasks/inject-data-application/define-command-argument-container/#running-a-command-in-a-shell

env: Option<Vec<EnvVar>>

List of environment variables to set in the container. Cannot be updated.

env_from: Option<Vec<EnvFromSource>>

List of sources to populate environment variables in the container. The keys defined within a source must be a C_IDENTIFIER. All invalid keys will be reported as an event when the container is starting. When a key exists in multiple sources, the value associated with the last source will take precedence. Values defined by an Env with a duplicate key will take precedence. Cannot be updated.

image: Option<String>

Container image name. More info: https://kubernetes.io/docs/concepts/containers/images

image_pull_policy: Option<String>

Image pull policy. One of Always, Never, IfNotPresent. Defaults to Always if :latest tag is specified, or IfNotPresent otherwise. Cannot be updated. More info: https://kubernetes.io/docs/concepts/containers/images#updating-images

lifecycle: Option<Lifecycle>

Lifecycle is not allowed for ephemeral containers.

liveness_probe: Option<Probe>

Probes are not allowed for ephemeral containers.

name: String

Name of the ephemeral container specified as a DNS_LABEL. This name must be unique among all containers, init containers and ephemeral containers.

ports: Option<Vec<ContainerPort>>

Ports are not allowed for ephemeral containers.

readiness_probe: Option<Probe>

Probes are not allowed for ephemeral containers.

resources: Option<ResourceRequirements>

Resources are not allowed for ephemeral containers. Ephemeral containers use spare resources already allocated to the pod.

security_context: Option<SecurityContext>

Optional: SecurityContext defines the security options the ephemeral container should be run with. If set, the fields of SecurityContext override the equivalent fields of PodSecurityContext.

startup_probe: Option<Probe>

Probes are not allowed for ephemeral containers.

stdin: Option<bool>

Whether this container should allocate a buffer for stdin in the container runtime. If this is not set, reads from stdin in the container will always result in EOF. Default is false.

stdin_once: Option<bool>

Whether the container runtime should close the stdin channel after it has been opened by a single attach. When stdin is true the stdin stream will remain open across multiple attach sessions. If stdinOnce is set to true, stdin is opened on container start, is empty until the first client attaches to stdin, and then remains open and accepts data until the client disconnects, at which time stdin is closed and remains closed until the container is restarted. If this flag is false, a container processes that reads from stdin will never receive an EOF. Default is false

target_container_name: Option<String>

If set, the name of the container from PodSpec that this ephemeral container targets. The ephemeral container will be run in the namespaces (IPC, PID, etc) of this container. If not set then the ephemeral container uses the namespaces configured in the Pod spec.

The container runtime must implement support for this feature. If the runtime does not support namespace targeting then the result of setting this field is undefined.

termination_message_path: Option<String>

Optional: Path at which the file to which the container’s termination message will be written is mounted into the container’s filesystem. Message written is intended to be brief final status, such as an assertion failure message. Will be truncated by the node if greater than 4096 bytes. The total message length across all containers will be limited to 12kb. Defaults to /dev/termination-log. Cannot be updated.

termination_message_policy: Option<String>

Indicate how the termination message should be populated. File will use the contents of terminationMessagePath to populate the container status message on both success and failure. FallbackToLogsOnError will use the last chunk of container log output if the termination message file is empty and the container exited with an error. The log output is limited to 2048 bytes or 80 lines, whichever is smaller. Defaults to File. Cannot be updated.

tty: Option<bool>

Whether this container should allocate a TTY for itself, also requires ‘stdin’ to be true. Default is false.

volume_devices: Option<Vec<VolumeDevice>>

volumeDevices is the list of block devices to be used by the container.

volume_mounts: Option<Vec<VolumeMount>>

Pod volumes to mount into the container’s filesystem. Subpath mounts are not allowed for ephemeral containers. Cannot be updated.

working_dir: Option<String>

Container’s working directory. If not specified, the container runtime’s default will be used, which might be configured in the container image. Cannot be updated.