Struct k8s_openapi::api::core::v1::Node

pub struct Node {
    pub metadata: ObjectMeta,
    pub spec: Option<NodeSpec>,
    pub status: Option<NodeStatus>,
}

Node is a worker node in Kubernetes. Each node will have a unique identifier in the cache (i.e. in etcd).

Fields
metadata: ObjectMeta

Standard object’s metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<NodeSpec>

Spec defines the behavior of a node. https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

status: Option<NodeStatus>

Most recently observed status of the node. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

Implementations

impl Node

pub fn connect_delete_node_proxy(
name: &str,
optional: ConnectDeleteNodeProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_delete_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectDeleteNodeProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_get_node_proxy(
name: &str,
optional: ConnectGetNodeProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_get_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectGetNodeProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_patch_node_proxy(
name: &str,
optional: ConnectPatchNodeProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_patch_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPatchNodeProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_post_node_proxy(
name: &str,
optional: ConnectPostNodeProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_post_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPostNodeProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_put_node_proxy(
name: &str,
optional: ConnectPutNodeProxyOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_put_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPutNodeProxyWithPathOptional<'_>
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn create_node(
body: &Node,
optional: CreateOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateResponse<Self>>), RequestError>

create a Node

Use the returned crate::ResponseBody<crate::CreateResponse<Self>> constructor, or crate::CreateResponse<Self> directly, to parse the HTTP response.

Arguments

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn delete_collection_node(
delete_optional: DeleteOptional<'_>,
list_optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<List<Self>>>), RequestError>

delete collection of Node

Use the returned crate::ResponseBody<crate::DeleteResponse<crate::List<Self>>> constructor, or crate::DeleteResponse<crate::List<Self>> directly, to parse the HTTP response.

Arguments

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Node

pub fn delete_node(
name: &str,
optional: DeleteOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteResponse<Self>>), RequestError>

delete a Node

Use the returned crate::ResponseBody<crate::DeleteResponse<Self>> constructor, or crate::DeleteResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn list_node(
optional: ListOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListResponse<Self>>), RequestError>

list or watch objects of kind Node

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<crate::ListResponse<Self>> constructor, or crate::ListResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn patch_node(
name: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update the specified Node

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn patch_node_status(
name: &str,
body: &Patch,
optional: PatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchResponse<Self>>), RequestError>

partially update status of the specified Node

Use the returned crate::ResponseBody<crate::PatchResponse<Self>> constructor, or crate::PatchResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn read_node(
name: &str,
optional: ReadNodeOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNodeResponse>), RequestError>

read the specified Node

Use the returned crate::ResponseBody<ReadNodeResponse> constructor, or ReadNodeResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn read_node_status(
name: &str,
optional: ReadNodeStatusOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNodeStatusResponse>), RequestError>

read status of the specified Node

Use the returned crate::ResponseBody<ReadNodeStatusResponse> constructor, or ReadNodeStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn replace_node(
name: &str,
body: &Node,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace the specified Node

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn replace_node_status(
name: &str,
body: &Node,
optional: ReplaceOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceResponse<Self>>), RequestError>

replace status of the specified Node

Use the returned crate::ResponseBody<crate::ReplaceResponse<Self>> constructor, or crate::ReplaceResponse<Self> directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn watch_node(
optional: WatchOptional<'_>
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchResponse<Self>>), RequestError>

list or watch objects of kind Node

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<crate::WatchResponse<Self>> constructor, or crate::WatchResponse<Self> directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.