Struct k8s_openapi::api::core::v1::Node

pub struct Node {
    pub metadata: Option<ObjectMeta>,
    pub spec: Option<NodeSpec>,
    pub status: Option<NodeStatus>,
}

Node is a worker node in Kubernetes. Each node will have a unique identifier in the cache (i.e. in etcd).

Fields
metadata: Option<ObjectMeta>

Standard object's metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata

spec: Option<NodeSpec>

Spec defines the behavior of a node. https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

status: Option<NodeStatus>

Most recently observed status of the node. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status

Methods

impl Node

pub fn connect_delete_node_proxy(
name: &str,
optional: ConnectDeleteNodeProxyOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_delete_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectDeleteNodeProxyWithPathOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect DELETE requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_get_node_proxy(
name: &str,
optional: ConnectGetNodeProxyOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_get_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectGetNodeProxyWithPathOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect GET requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_patch_node_proxy(
name: &str,
optional: ConnectPatchNodeProxyOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_patch_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPatchNodeProxyWithPathOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect PATCH requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_post_node_proxy(
name: &str,
optional: ConnectPostNodeProxyOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_post_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPostNodeProxyWithPathOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect POST requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_put_node_proxy(
name: &str,
optional: ConnectPutNodeProxyOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn connect_put_node_proxy_with_path(
name: &str,
path: &str,
optional: ConnectPutNodeProxyWithPathOptional
) -> Result<Request<Vec<u8>>, RequestError>

connect PUT requests to proxy of Node

Arguments

name

name of the NodeProxyOptions

path

path to the resource

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn create_node(
body: &Node,
optional: CreateNodeOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateNodeResponse>), RequestError>

create a Node

Use the returned crate::ResponseBody<CreateNodeResponse> constructor, or CreateNodeResponse directly, to parse the HTTP response.

Arguments

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn delete_collection_node(
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteCollectionNodeResponse>), RequestError>

delete collection of Node

Use the returned crate::ResponseBody<DeleteCollectionNodeResponse> constructor, or DeleteCollectionNodeResponse directly, to parse the HTTP response.

Arguments

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Node

pub fn delete_node(
name: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteNodeResponse>), RequestError>

delete a Node

Use the returned crate::ResponseBody<DeleteNodeResponse> constructor, or DeleteNodeResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn list_node(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListNodeResponse>), RequestError>

list or watch objects of kind Node

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListNodeResponse> constructor, or ListNodeResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn patch_node(
name: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNodeResponse>), RequestError>

partially update the specified Node

Use the returned crate::ResponseBody<PatchNodeResponse> constructor, or PatchNodeResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn patch_node_status(
name: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNodeStatusResponse>), RequestError>

partially update status of the specified Node

Use the returned crate::ResponseBody<PatchNodeStatusResponse> constructor, or PatchNodeStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn read_node(
name: &str,
optional: ReadNodeOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNodeResponse>), RequestError>

read the specified Node

Use the returned crate::ResponseBody<ReadNodeResponse> constructor, or ReadNodeResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn read_node_status(
name: &str,
optional: ReadNodeStatusOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNodeStatusResponse>), RequestError>

read status of the specified Node

Use the returned crate::ResponseBody<ReadNodeStatusResponse> constructor, or ReadNodeStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn replace_node(
name: &str,
body: &Node,
optional: ReplaceNodeOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNodeResponse>), RequestError>

replace the specified Node

Use the returned crate::ResponseBody<ReplaceNodeResponse> constructor, or ReplaceNodeResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn replace_node_status(
name: &str,
body: &Node,
optional: ReplaceNodeStatusOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNodeStatusResponse>), RequestError>

replace status of the specified Node

Use the returned crate::ResponseBody<ReplaceNodeStatusResponse> constructor, or ReplaceNodeStatusResponse directly, to parse the HTTP response.

Arguments

name

name of the Node

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Node

pub fn watch_node(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchNodeResponse>), RequestError>

list or watch objects of kind Node

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchNodeResponse> constructor, or WatchNodeResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.