Struct k8s_openapi::api::core::v1::PersistentVolumeSpec

pub struct PersistentVolumeSpec {

    pub access_modes: Option<Vec<String>>,
    pub aws_elastic_block_store: Option<AWSElasticBlockStoreVolumeSource>,
    pub azure_disk: Option<AzureDiskVolumeSource>,
    pub azure_file: Option<AzureFilePersistentVolumeSource>,
    pub capacity: Option<BTreeMap<String, Quantity>>,
    pub cephfs: Option<CephFSPersistentVolumeSource>,
    pub cinder: Option<CinderPersistentVolumeSource>,
    pub claim_ref: Option<ObjectReference>,
    pub csi: Option<CSIPersistentVolumeSource>,
    pub fc: Option<FCVolumeSource>,
    pub flex_volume: Option<FlexPersistentVolumeSource>,
    pub flocker: Option<FlockerVolumeSource>,
    pub gce_persistent_disk: Option<GCEPersistentDiskVolumeSource>,
    pub glusterfs: Option<GlusterfsPersistentVolumeSource>,
    pub host_path: Option<HostPathVolumeSource>,
    pub iscsi: Option<ISCSIPersistentVolumeSource>,
    pub local: Option<LocalVolumeSource>,
    pub mount_options: Option<Vec<String>>,
    pub nfs: Option<NFSVolumeSource>,
    pub node_affinity: Option<VolumeNodeAffinity>,
    pub persistent_volume_reclaim_policy: Option<String>,
    pub photon_persistent_disk: Option<PhotonPersistentDiskVolumeSource>,
    pub portworx_volume: Option<PortworxVolumeSource>,
    pub quobyte: Option<QuobyteVolumeSource>,
    pub rbd: Option<RBDPersistentVolumeSource>,
    pub scale_io: Option<ScaleIOPersistentVolumeSource>,
    pub storage_class_name: Option<String>,
    pub storageos: Option<StorageOSPersistentVolumeSource>,
    pub volume_mode: Option<String>,
    pub vsphere_volume: Option<VsphereVirtualDiskVolumeSource>,

}

PersistentVolumeSpec is the specification of a persistent volume.

Fields
access_modes: Option<Vec<String>>

AccessModes contains all ways the volume can be mounted. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#access-modes

aws_elastic_block_store: Option<AWSElasticBlockStoreVolumeSource>

AWSElasticBlockStore represents an AWS Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. More info: https://kubernetes.io/docs/concepts/storage/volumes#awselasticblockstore

azure_disk: Option<AzureDiskVolumeSource>

AzureDisk represents an Azure Data Disk mount on the host and bind mount to the pod.

azure_file: Option<AzureFilePersistentVolumeSource>

AzureFile represents an Azure File Service mount on the host and bind mount to the pod.

capacity: Option<BTreeMap<String, Quantity>>

A description of the persistent volume’s resources and capacity. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#capacity

cephfs: Option<CephFSPersistentVolumeSource>

CephFS represents a Ceph FS mount on the host that shares a pod’s lifetime

cinder: Option<CinderPersistentVolumeSource>

Cinder represents a cinder volume attached and mounted on kubelets host machine. More info: https://examples.k8s.io/mysql-cinder-pd/README.md

claim_ref: Option<ObjectReference>

ClaimRef is part of a bi-directional binding between PersistentVolume and PersistentVolumeClaim. Expected to be non-nil when bound. claim.VolumeName is the authoritative bind between PV and PVC. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#binding

csi: Option<CSIPersistentVolumeSource>

CSI represents storage that is handled by an external CSI driver (Beta feature).

fc: Option<FCVolumeSource>

FC represents a Fibre Channel resource that is attached to a kubelet’s host machine and then exposed to the pod.

flex_volume: Option<FlexPersistentVolumeSource>

FlexVolume represents a generic volume resource that is provisioned/attached using an exec based plugin.

flocker: Option<FlockerVolumeSource>

Flocker represents a Flocker volume attached to a kubelet’s host machine and exposed to the pod for its usage. This depends on the Flocker control service being running

gce_persistent_disk: Option<GCEPersistentDiskVolumeSource>

GCEPersistentDisk represents a GCE Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. Provisioned by an admin. More info: https://kubernetes.io/docs/concepts/storage/volumes#gcepersistentdisk

glusterfs: Option<GlusterfsPersistentVolumeSource>

Glusterfs represents a Glusterfs volume that is attached to a host and exposed to the pod. Provisioned by an admin. More info: https://examples.k8s.io/volumes/glusterfs/README.md

host_path: Option<HostPathVolumeSource>

HostPath represents a directory on the host. Provisioned by a developer or tester. This is useful for single-node development and testing only! On-host storage is not supported in any way and WILL NOT WORK in a multi-node cluster. More info: https://kubernetes.io/docs/concepts/storage/volumes#hostpath

iscsi: Option<ISCSIPersistentVolumeSource>

ISCSI represents an ISCSI Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. Provisioned by an admin.

local: Option<LocalVolumeSource>

Local represents directly-attached storage with node affinity

mount_options: Option<Vec<String>>

A list of mount options, e.g. [“ro”, “soft”]. Not validated - mount will simply fail if one is invalid. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes/#mount-options

nfs: Option<NFSVolumeSource>

NFS represents an NFS mount on the host. Provisioned by an admin. More info: https://kubernetes.io/docs/concepts/storage/volumes#nfs

node_affinity: Option<VolumeNodeAffinity>

NodeAffinity defines constraints that limit what nodes this volume can be accessed from. This field influences the scheduling of pods that use this volume.

persistent_volume_reclaim_policy: Option<String>

What happens to a persistent volume when released from its claim. Valid options are Retain (default for manually created PersistentVolumes), Delete (default for dynamically provisioned PersistentVolumes), and Recycle (deprecated). Recycle must be supported by the volume plugin underlying this PersistentVolume. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#reclaiming

photon_persistent_disk: Option<PhotonPersistentDiskVolumeSource>

PhotonPersistentDisk represents a PhotonController persistent disk attached and mounted on kubelets host machine

portworx_volume: Option<PortworxVolumeSource>

PortworxVolume represents a portworx volume attached and mounted on kubelets host machine

quobyte: Option<QuobyteVolumeSource>

Quobyte represents a Quobyte mount on the host that shares a pod’s lifetime

rbd: Option<RBDPersistentVolumeSource>

RBD represents a Rados Block Device mount on the host that shares a pod’s lifetime. More info: https://examples.k8s.io/volumes/rbd/README.md

scale_io: Option<ScaleIOPersistentVolumeSource>

ScaleIO represents a ScaleIO persistent volume attached and mounted on Kubernetes nodes.

storage_class_name: Option<String>

Name of StorageClass to which this persistent volume belongs. Empty value means that this volume does not belong to any StorageClass.

storageos: Option<StorageOSPersistentVolumeSource>

StorageOS represents a StorageOS volume that is attached to the kubelet’s host machine and mounted into the pod More info: https://examples.k8s.io/volumes/storageos/README.md

volume_mode: Option<String>

volumeMode defines if a volume is intended to be used with a formatted filesystem or to remain in raw block state. Value of Filesystem is implied when not included in spec.

vsphere_volume: Option<VsphereVirtualDiskVolumeSource>

VsphereVolume represents a vSphere volume attached and mounted on kubelets host machine