k8s_openapi::api::core::v1

Struct Volume

pub struct Volume {

    pub aws_elastic_block_store: Option<AWSElasticBlockStoreVolumeSource>,
    pub azure_disk: Option<AzureDiskVolumeSource>,
    pub azure_file: Option<AzureFileVolumeSource>,
    pub cephfs: Option<CephFSVolumeSource>,
    pub cinder: Option<CinderVolumeSource>,
    pub config_map: Option<ConfigMapVolumeSource>,
    pub csi: Option<CSIVolumeSource>,
    pub downward_api: Option<DownwardAPIVolumeSource>,
    pub empty_dir: Option<EmptyDirVolumeSource>,
    pub ephemeral: Option<EphemeralVolumeSource>,
    pub fc: Option<FCVolumeSource>,
    pub flex_volume: Option<FlexVolumeSource>,
    pub flocker: Option<FlockerVolumeSource>,
    pub gce_persistent_disk: Option<GCEPersistentDiskVolumeSource>,
    pub git_repo: Option<GitRepoVolumeSource>,
    pub glusterfs: Option<GlusterfsVolumeSource>,
    pub host_path: Option<HostPathVolumeSource>,
    pub image: Option<ImageVolumeSource>,
    pub iscsi: Option<ISCSIVolumeSource>,
    pub name: String,
    pub nfs: Option<NFSVolumeSource>,
    pub persistent_volume_claim: Option<PersistentVolumeClaimVolumeSource>,
    pub photon_persistent_disk: Option<PhotonPersistentDiskVolumeSource>,
    pub portworx_volume: Option<PortworxVolumeSource>,
    pub projected: Option<ProjectedVolumeSource>,
    pub quobyte: Option<QuobyteVolumeSource>,
    pub rbd: Option<RBDVolumeSource>,
    pub scale_io: Option<ScaleIOVolumeSource>,
    pub secret: Option<SecretVolumeSource>,
    pub storageos: Option<StorageOSVolumeSource>,
    pub vsphere_volume: Option<VsphereVirtualDiskVolumeSource>,

}

Volume represents a named volume in a pod that may be accessed by any container in the pod.

Fields
aws_elastic_block_store: Option<AWSElasticBlockStoreVolumeSource>

awsElasticBlockStore represents an AWS Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. Deprecated: AWSElasticBlockStore is deprecated. All operations for the in-tree awsElasticBlockStore type are redirected to the ebs.csi.aws.com CSI driver. More info: https://kubernetes.io/docs/concepts/storage/volumes#awselasticblockstore

azure_disk: Option<AzureDiskVolumeSource>

azureDisk represents an Azure Data Disk mount on the host and bind mount to the pod. Deprecated: AzureDisk is deprecated. All operations for the in-tree azureDisk type are redirected to the disk.csi.azure.com CSI driver.

azure_file: Option<AzureFileVolumeSource>

azureFile represents an Azure File Service mount on the host and bind mount to the pod. Deprecated: AzureFile is deprecated. All operations for the in-tree azureFile type are redirected to the file.csi.azure.com CSI driver.

cephfs: Option<CephFSVolumeSource>

cephFS represents a Ceph FS mount on the host that shares a pod’s lifetime. Deprecated: CephFS is deprecated and the in-tree cephfs type is no longer supported.

cinder: Option<CinderVolumeSource>

cinder represents a cinder volume attached and mounted on kubelets host machine. Deprecated: Cinder is deprecated. All operations for the in-tree cinder type are redirected to the cinder.csi.openstack.org CSI driver. More info: https://examples.k8s.io/mysql-cinder-pd/README.md

config_map: Option<ConfigMapVolumeSource>

configMap represents a configMap that should populate this volume

csi: Option<CSIVolumeSource>

csi (Container Storage Interface) represents ephemeral storage that is handled by certain external CSI drivers.

downward_api: Option<DownwardAPIVolumeSource>

downwardAPI represents downward API about the pod that should populate this volume

empty_dir: Option<EmptyDirVolumeSource>

emptyDir represents a temporary directory that shares a pod’s lifetime. More info: https://kubernetes.io/docs/concepts/storage/volumes#emptydir

ephemeral: Option<EphemeralVolumeSource>

ephemeral represents a volume that is handled by a cluster storage driver. The volume’s lifecycle is tied to the pod that defines it - it will be created before the pod starts, and deleted when the pod is removed.

Use this if: a) the volume is only needed while the pod runs, b) features of normal volumes like restoring from snapshot or capacity tracking are needed, c) the storage driver is specified through a storage class, and d) the storage driver supports dynamic volume provisioning through a PersistentVolumeClaim (see EphemeralVolumeSource for more information on the connection between this volume type and PersistentVolumeClaim).

Use PersistentVolumeClaim or one of the vendor-specific APIs for volumes that persist for longer than the lifecycle of an individual pod.

Use CSI for light-weight local ephemeral volumes if the CSI driver is meant to be used that way - see the documentation of the driver for more information.

A pod can use both types of ephemeral volumes and persistent volumes at the same time.

fc: Option<FCVolumeSource>

fc represents a Fibre Channel resource that is attached to a kubelet’s host machine and then exposed to the pod.

flex_volume: Option<FlexVolumeSource>

flexVolume represents a generic volume resource that is provisioned/attached using an exec based plugin. Deprecated: FlexVolume is deprecated. Consider using a CSIDriver instead.

flocker: Option<FlockerVolumeSource>

flocker represents a Flocker volume attached to a kubelet’s host machine. This depends on the Flocker control service being running. Deprecated: Flocker is deprecated and the in-tree flocker type is no longer supported.

gce_persistent_disk: Option<GCEPersistentDiskVolumeSource>

gcePersistentDisk represents a GCE Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. Deprecated: GCEPersistentDisk is deprecated. All operations for the in-tree gcePersistentDisk type are redirected to the pd.csi.storage.gke.io CSI driver. More info: https://kubernetes.io/docs/concepts/storage/volumes#gcepersistentdisk

git_repo: Option<GitRepoVolumeSource>

gitRepo represents a git repository at a particular revision. Deprecated: GitRepo is deprecated. To provision a container with a git repo, mount an EmptyDir into an InitContainer that clones the repo using git, then mount the EmptyDir into the Pod’s container.

glusterfs: Option<GlusterfsVolumeSource>

glusterfs represents a Glusterfs mount on the host that shares a pod’s lifetime. Deprecated: Glusterfs is deprecated and the in-tree glusterfs type is no longer supported.

host_path: Option<HostPathVolumeSource>

hostPath represents a pre-existing file or directory on the host machine that is directly exposed to the container. This is generally used for system agents or other privileged things that are allowed to see the host machine. Most containers will NOT need this. More info: https://kubernetes.io/docs/concepts/storage/volumes#hostpath

image: Option<ImageVolumeSource>

image represents an OCI object (a container image or artifact) pulled and mounted on the kubelet’s host machine. The volume is resolved at pod startup depending on which PullPolicy value is provided:

Always: the kubelet always attempts to pull the reference. Container creation will fail If the pull fails. - Never: the kubelet never pulls the reference and only uses a local image or artifact. Container creation will fail if the reference isn’t present. - IfNotPresent: the kubelet pulls if the reference isn’t already present on disk. Container creation will fail if the reference isn’t present and the pull fails.

The volume gets re-resolved if the pod gets deleted and recreated, which means that new remote content will become available on pod recreation. A failure to resolve or pull the image during pod startup will block containers from starting and may add significant latency. Failures will be retried using normal volume backoff and will be reported on the pod reason and message. The types of objects that may be mounted by this volume are defined by the container runtime implementation on a host machine and at minimum must include all valid types supported by the container image field. The OCI object gets mounted in a single directory (spec.containers[].volumeMounts.mountPath) by merging the manifest layers in the same way as for container images. The volume will be mounted read-only (ro) and non-executable files (noexec). Sub path mounts for containers are not supported (spec.containers[].volumeMounts.subpath) before 1.33. The field spec.securityContext.fsGroupChangePolicy has no effect on this volume type.

iscsi: Option<ISCSIVolumeSource>

iscsi represents an ISCSI Disk resource that is attached to a kubelet’s host machine and then exposed to the pod. More info: https://kubernetes.io/docs/concepts/storage/volumes/#iscsi

name: String

name of the volume. Must be a DNS_LABEL and unique within the pod. More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names

nfs: Option<NFSVolumeSource>

nfs represents an NFS mount on the host that shares a pod’s lifetime More info: https://kubernetes.io/docs/concepts/storage/volumes#nfs

persistent_volume_claim: Option<PersistentVolumeClaimVolumeSource>

persistentVolumeClaimVolumeSource represents a reference to a PersistentVolumeClaim in the same namespace. More info: https://kubernetes.io/docs/concepts/storage/persistent-volumes#persistentvolumeclaims

photon_persistent_disk: Option<PhotonPersistentDiskVolumeSource>

photonPersistentDisk represents a PhotonController persistent disk attached and mounted on kubelets host machine. Deprecated: PhotonPersistentDisk is deprecated and the in-tree photonPersistentDisk type is no longer supported.

portworx_volume: Option<PortworxVolumeSource>

portworxVolume represents a portworx volume attached and mounted on kubelets host machine. Deprecated: PortworxVolume is deprecated. All operations for the in-tree portworxVolume type are redirected to the pxd.portworx.com CSI driver when the CSIMigrationPortworx feature-gate is on.

projected: Option<ProjectedVolumeSource>

projected items for all in one resources secrets, configmaps, and downward API

quobyte: Option<QuobyteVolumeSource>

quobyte represents a Quobyte mount on the host that shares a pod’s lifetime. Deprecated: Quobyte is deprecated and the in-tree quobyte type is no longer supported.

rbd: Option<RBDVolumeSource>

rbd represents a Rados Block Device mount on the host that shares a pod’s lifetime. Deprecated: RBD is deprecated and the in-tree rbd type is no longer supported.

scale_io: Option<ScaleIOVolumeSource>

scaleIO represents a ScaleIO persistent volume attached and mounted on Kubernetes nodes. Deprecated: ScaleIO is deprecated and the in-tree scaleIO type is no longer supported.

secret: Option<SecretVolumeSource>

secret represents a secret that should populate this volume. More info: https://kubernetes.io/docs/concepts/storage/volumes#secret

storageos: Option<StorageOSVolumeSource>

storageOS represents a StorageOS volume attached and mounted on Kubernetes nodes. Deprecated: StorageOS is deprecated and the in-tree storageos type is no longer supported.

vsphere_volume: Option<VsphereVirtualDiskVolumeSource>

vsphereVolume represents a vSphere volume attached and mounted on kubelets host machine. Deprecated: VsphereVolume is deprecated. All operations for the in-tree vsphereVolume type are redirected to the csi.vsphere.vmware.com CSI driver.