k8s_openapi::api::discovery

Module v1

Structs

Endpoint

Endpoint represents a single logical “backend” implementing a service.

EndpointConditions

EndpointConditions represents the current condition of an endpoint.

EndpointHints

EndpointHints provides hints describing how an endpoint should be consumed.

EndpointPort

EndpointPort represents a Port used by an EndpointSlice

EndpointSlice

EndpointSlice represents a set of service endpoints. Most EndpointSlices are created by the EndpointSlice controller to represent the Pods selected by Service objects. For a given service there may be multiple EndpointSlice objects which must be joined to produce the full set of endpoints; you can find all of the slices for a given service by listing EndpointSlices in the service’s namespace whose kubernetes.io/service-name label contains the service’s name.

ForNode

ForNode provides information about which nodes should consume this endpoint.

ForZone

ForZone provides information about which zones should consume this endpoint.