Module k8s_openapi::api::discovery::v1

Structs

Endpoint

Endpoint represents a single logical “backend” implementing a service.

EndpointConditions

EndpointConditions represents the current condition of an endpoint.

EndpointHints

EndpointHints provides hints describing how an endpoint should be consumed.

EndpointPort

EndpointPort represents a Port used by an EndpointSlice

EndpointSlice

EndpointSlice represents a subset of the endpoints that implement a service. For a given service there may be multiple EndpointSlice objects, selected by labels, which must be joined to produce the full set of endpoints.

ForZone

ForZone provides information about which zones should consume this endpoint.

ReadNamespacedEndpointSliceOptional

Optional parameters of EndpointSlice::read_namespaced_endpoint_slice

Enums

ReadNamespacedEndpointSliceResponse

Use <ReadNamespacedEndpointSliceResponse as Response>::try_from_parts to parse the HTTP response body of EndpointSlice::read_namespaced_endpoint_slice