Struct k8s_openapi::api::events::v1beta1::Event

pub struct Event {
    pub action: Option<String>,
    pub deprecated_count: Option<i32>,
    pub deprecated_first_timestamp: Option<Time>,
    pub deprecated_last_timestamp: Option<Time>,
    pub deprecated_source: Option<EventSource>,
    pub event_time: MicroTime,
    pub metadata: Option<ObjectMeta>,
    pub note: Option<String>,
    pub reason: Option<String>,
    pub regarding: Option<ObjectReference>,
    pub related: Option<ObjectReference>,
    pub reporting_controller: Option<String>,
    pub reporting_instance: Option<String>,
    pub series: Option<EventSeries>,
    pub type_: Option<String>,
}

Event is a report of an event somewhere in the cluster. It generally denotes some state change in the system.

Fields
action: Option<String>

What action was taken/failed regarding to the regarding object.

deprecated_count: Option<i32>

Deprecated field assuring backward compatibility with core.v1 Event type

deprecated_first_timestamp: Option<Time>

Deprecated field assuring backward compatibility with core.v1 Event type

deprecated_last_timestamp: Option<Time>

Deprecated field assuring backward compatibility with core.v1 Event type

deprecated_source: Option<EventSource>

Deprecated field assuring backward compatibility with core.v1 Event type

event_time: MicroTime

Required. Time when this Event was first observed.

metadata: Option<ObjectMeta>note: Option<String>

Optional. A human-readable description of the status of this operation. Maximal length of the note is 1kB, but libraries should be prepared to handle values up to 64kB.

reason: Option<String>

Why the action was taken.

regarding: Option<ObjectReference>

The object this Event is about. In most cases it's an Object reporting controller implements. E.g. ReplicaSetController implements ReplicaSets and this event is emitted because it acts on some changes in a ReplicaSet object.

related: Option<ObjectReference>

Optional secondary object for more complex actions. E.g. when regarding object triggers a creation or deletion of related object.

reporting_controller: Option<String>

Name of the controller that emitted this Event, e.g. kubernetes.io/kubelet.

reporting_instance: Option<String>

ID of the controller instance, e.g. kubelet-xyzf.

series: Option<EventSeries>

Data about the Event series this event represents or nil if it's a singleton Event.

type_: Option<String>

Type of this event (Normal, Warning), new types could be added in the future.

Methods

impl Event

pub fn create_namespaced_event(
namespace: &str,
body: &Event,
optional: CreateNamespacedEventOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<CreateNamespacedEventResponse>), RequestError>

create an Event

Use the returned crate::ResponseBody<CreateNamespacedEventResponse> constructor, or CreateNamespacedEventResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn delete_collection_namespaced_event(
namespace: &str,
delete_optional: DeleteOptional,
list_optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteCollectionNamespacedEventResponse>), RequestError>

delete collection of Event

Use the returned crate::ResponseBody<DeleteCollectionNamespacedEventResponse> constructor, or DeleteCollectionNamespacedEventResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

delete_optional

Delete options. Use Default::default() to not pass any.

list_optional

List options. Use Default::default() to not pass any.

impl Event

pub fn delete_namespaced_event(
name: &str,
namespace: &str,
optional: DeleteOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<DeleteNamespacedEventResponse>), RequestError>

delete an Event

Use the returned crate::ResponseBody<DeleteNamespacedEventResponse> constructor, or DeleteNamespacedEventResponse directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn list_event_for_all_namespaces(
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListEventForAllNamespacesResponse>), RequestError>

list or watch objects of kind Event

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListEventForAllNamespacesResponse> constructor, or ListEventForAllNamespacesResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn list_namespaced_event(
namespace: &str,
optional: ListOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ListNamespacedEventResponse>), RequestError>

list or watch objects of kind Event

This operation only supports listing all items of this type.

Use the returned crate::ResponseBody<ListNamespacedEventResponse> constructor, or ListNamespacedEventResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn patch_namespaced_event(
name: &str,
namespace: &str,
body: &Patch,
optional: PatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<PatchNamespacedEventResponse>), RequestError>

partially update the specified Event

Use the returned crate::ResponseBody<PatchNamespacedEventResponse> constructor, or PatchNamespacedEventResponse directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn read_namespaced_event(
name: &str,
namespace: &str,
optional: ReadNamespacedEventOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReadNamespacedEventResponse>), RequestError>

read the specified Event

Use the returned crate::ResponseBody<ReadNamespacedEventResponse> constructor, or ReadNamespacedEventResponse directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn replace_namespaced_event(
name: &str,
namespace: &str,
body: &Event,
optional: ReplaceNamespacedEventOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<ReplaceNamespacedEventResponse>), RequestError>

replace the specified Event

Use the returned crate::ResponseBody<ReplaceNamespacedEventResponse> constructor, or ReplaceNamespacedEventResponse directly, to parse the HTTP response.

Arguments

name

name of the Event

namespace

object name and auth scope, such as for teams and projects

body

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn watch_event_for_all_namespaces(
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchEventForAllNamespacesResponse>), RequestError>

list or watch objects of kind Event

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchEventForAllNamespacesResponse> constructor, or WatchEventForAllNamespacesResponse directly, to parse the HTTP response.

Arguments

optional

Optional parameters. Use Default::default() to not pass any.

impl Event

pub fn watch_namespaced_event(
namespace: &str,
optional: WatchOptional
) -> Result<(Request<Vec<u8>>, fn(_: StatusCode) -> ResponseBody<WatchNamespacedEventResponse>), RequestError>

list or watch objects of kind Event

This operation only supports watching one item, or a list of items, of this type for changes.

Use the returned crate::ResponseBody<WatchNamespacedEventResponse> constructor, or WatchNamespacedEventResponse directly, to parse the HTTP response.

Arguments

namespace

object name and auth scope, such as for teams and projects

optional

Optional parameters. Use Default::default() to not pass any.