Module k8s_openapi::api::networking::v1

Structs

IPBlock
IPBlock describes a particular CIDR (Ex. "192.168.1.1/24","2001:db9::/64") that is allowed to the pods matched by a NetworkPolicySpec's podSelector. The except entry describes CIDRs that should not be included within this rule.

NetworkPolicy
NetworkPolicy describes what network traffic is allowed for a set of Pods

NetworkPolicyEgressRule
NetworkPolicyEgressRule describes a particular set of traffic that is allowed out of pods matched by a NetworkPolicySpec's podSelector. The traffic must match both ports and to. This type is beta-level in 1.8

NetworkPolicyIngressRule
NetworkPolicyIngressRule describes a particular set of traffic that is allowed to the pods matched by a NetworkPolicySpec's podSelector. The traffic must match both ports and from.

NetworkPolicyPeer
NetworkPolicyPeer describes a peer to allow traffic from. Only certain combinations of fields are allowed

NetworkPolicyPort
NetworkPolicyPort describes a port to allow traffic on

NetworkPolicySpec
NetworkPolicySpec provides the specification of a NetworkPolicy

ReadNamespacedNetworkPolicyOptional
Optional parameters of NetworkPolicy::read_namespaced_network_policy

Enums

ReadNamespacedNetworkPolicyResponse
Use <ReadNamespacedNetworkPolicyResponse as Response>::try_from_parts to parse the HTTP response body of NetworkPolicy::read_namespaced_network_policy