Module k8s_openapi::api::rbac::v1

Structs

AggregationRule

AggregationRule describes how to locate ClusterRoles to aggregate into the ClusterRole

ClusterRole

ClusterRole is a cluster level, logical grouping of PolicyRules that can be referenced as a unit by a RoleBinding or ClusterRoleBinding.

ClusterRoleBinding

ClusterRoleBinding references a ClusterRole, but not contain it. It can reference a ClusterRole in the global namespace, and adds who information via Subject.

PolicyRule

PolicyRule holds information that describes a policy rule, but does not contain information about who the rule applies to or which namespace the rule applies to.

ReadClusterRoleBindingOptional

Optional parameters of ClusterRoleBinding::read_cluster_role_binding

ReadClusterRoleOptional

Optional parameters of ClusterRole::read_cluster_role

ReadNamespacedRoleBindingOptional

Optional parameters of RoleBinding::read_namespaced_role_binding

ReadNamespacedRoleOptional

Optional parameters of Role::read_namespaced_role

Role

Role is a namespaced, logical grouping of PolicyRules that can be referenced as a unit by a RoleBinding.

RoleBinding

RoleBinding references a role, but does not contain it. It can reference a Role in the same namespace or a ClusterRole in the global namespace. It adds who information via Subjects and namespace information by which namespace it exists in. RoleBindings in a given namespace only have effect in that namespace.

RoleRef

RoleRef contains information that points to the role being used

Subject

Subject contains a reference to the object or user identities a role binding applies to. This can either hold a direct API object reference, or a value for non-objects such as user and group names.

Enums

ReadClusterRoleBindingResponse

Use <ReadClusterRoleBindingResponse as Response>::try_from_parts to parse the HTTP response body of ClusterRoleBinding::read_cluster_role_binding

ReadClusterRoleResponse

Use <ReadClusterRoleResponse as Response>::try_from_parts to parse the HTTP response body of ClusterRole::read_cluster_role

ReadNamespacedRoleBindingResponse

Use <ReadNamespacedRoleBindingResponse as Response>::try_from_parts to parse the HTTP response body of RoleBinding::read_namespaced_role_binding

ReadNamespacedRoleResponse

Use <ReadNamespacedRoleResponse as Response>::try_from_parts to parse the HTTP response body of Role::read_namespaced_role