Struct k8s_openapi::apiextensions_apiserver::pkg::apis::apiextensions::v1::JSONSchemaProps

pub struct JSONSchemaProps {
    pub ref_path: Option<String>,
    pub schema: Option<String>,
    pub additional_items: Option<JSONSchemaPropsOrBool>,
    pub additional_properties: Option<JSONSchemaPropsOrBool>,
    pub all_of: Option<Vec<JSONSchemaProps>>,
    pub any_of: Option<Vec<JSONSchemaProps>>,
    pub default: Option<JSON>,
    pub definitions: Option<BTreeMap<String, JSONSchemaProps>>,
    pub dependencies: Option<BTreeMap<String, JSONSchemaPropsOrStringArray>>,
    pub description: Option<String>,
    pub enum_: Option<Vec<JSON>>,
    pub example: Option<JSON>,
    pub exclusive_maximum: Option<bool>,
    pub exclusive_minimum: Option<bool>,
    pub external_docs: Option<ExternalDocumentation>,
    pub format: Option<String>,
    pub id: Option<String>,
    pub items: Option<JSONSchemaPropsOrArray>,
    pub max_items: Option<i64>,
    pub max_length: Option<i64>,
    pub max_properties: Option<i64>,
    pub maximum: Option<f64>,
    pub min_items: Option<i64>,
    pub min_length: Option<i64>,
    pub min_properties: Option<i64>,
    pub minimum: Option<f64>,
    pub multiple_of: Option<f64>,
    pub not: Option<Box<JSONSchemaProps>>,
    pub nullable: Option<bool>,
    pub one_of: Option<Vec<JSONSchemaProps>>,
    pub pattern: Option<String>,
    pub pattern_properties: Option<BTreeMap<String, JSONSchemaProps>>,
    pub properties: Option<BTreeMap<String, JSONSchemaProps>>,
    pub required: Option<Vec<String>>,
    pub title: Option<String>,
    pub type_: Option<String>,
    pub unique_items: Option<bool>,
    pub x_kubernetes_embedded_resource: Option<bool>,
    pub x_kubernetes_int_or_string: Option<bool>,
    pub x_kubernetes_list_map_keys: Option<Vec<String>>,
    pub x_kubernetes_list_type: Option<String>,
    pub x_kubernetes_map_type: Option<String>,
    pub x_kubernetes_preserve_unknown_fields: Option<bool>,
}

JSONSchemaProps is a JSON-Schema following Specification Draft 4 (http://json-schema.org/).

Fields
ref_path: Option<String>schema: Option<String>additional_items: Option<JSONSchemaPropsOrBool>additional_properties: Option<JSONSchemaPropsOrBool>all_of: Option<Vec<JSONSchemaProps>>any_of: Option<Vec<JSONSchemaProps>>default: Option<JSON>

default is a default value for undefined object fields. Defaulting is a beta feature under the CustomResourceDefaulting feature gate. Defaulting requires spec.preserveUnknownFields to be false.

definitions: Option<BTreeMap<String, JSONSchemaProps>>dependencies: Option<BTreeMap<String, JSONSchemaPropsOrStringArray>>description: Option<String>enum_: Option<Vec<JSON>>example: Option<JSON>exclusive_maximum: Option<bool>exclusive_minimum: Option<bool>external_docs: Option<ExternalDocumentation>format: Option<String>

format is an OpenAPI v3 format string. Unknown formats are ignored. The following formats are validated:

bsonobjectid: a bson object ID, i.e. a 24 characters hex string - uri: an URI as parsed by Golang net/url.ParseRequestURI - email: an email address as parsed by Golang net/mail.ParseAddress - hostname: a valid representation for an Internet host name, as defined by RFC 1034, section 3.1 [RFC1034]. - ipv4: an IPv4 IP as parsed by Golang net.ParseIP - ipv6: an IPv6 IP as parsed by Golang net.ParseIP - cidr: a CIDR as parsed by Golang net.ParseCIDR - mac: a MAC address as parsed by Golang net.ParseMAC - uuid: an UUID that allows uppercase defined by the regex (?i)^[0-9a-f]{8}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{12}$ - uuid3: an UUID3 that allows uppercase defined by the regex (?i)^[0-9a-f]{8}-?[0-9a-f]{4}-?3[0-9a-f]{3}-?[0-9a-f]{4}-?[0-9a-f]{12}$ - uuid4: an UUID4 that allows uppercase defined by the regex (?i)^[0-9a-f]{8}-?[0-9a-f]{4}-?4[0-9a-f]{3}-?[89ab][0-9a-f]{3}-?[0-9a-f]{12}$ - uuid5: an UUID5 that allows uppercase defined by the regex (?i)^[0-9a-f]{8}-?[0-9a-f]{4}-?5[0-9a-f]{3}-?[89ab][0-9a-f]{3}-?[0-9a-f]{12}$ - isbn: an ISBN10 or ISBN13 number string like "0321751043" or "978-0321751041" - isbn10: an ISBN10 number string like "0321751043" - isbn13: an ISBN13 number string like "978-0321751041" - creditcard: a credit card number defined by the regex ^(?:4[0-9]{12}(?:[0-9]{3})?|5[1-5][0-9]{14}|6(?:011|5[0-9][0-9])[0-9]{12}|3[47][0-9]{13}|3(?:0[0-5]|[68][0-9])[0-9]{11}|(?:2131|1800|35\d{3})\d{11})$ with any non digit characters mixed in - ssn: a U.S. social security number following the regex ^\d{3}[- ]?\d{2}[- ]?\d{4}$ - hexcolor: an hexadecimal color code like "#FFFFFF: following the regex ^#?([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$ - rgbcolor: an RGB color code like rgb like "rgb(255,255,2559" - byte: base64 encoded binary data - password: any kind of string - date: a date string like "2006-01-02" as defined by full-date in RFC3339 - duration: a duration string like "22 ns" as parsed by Golang time.ParseDuration or compatible with Scala duration format - datetime: a date time string like "2014-12-15T19:30:20.000Z" as defined by date-time in RFC3339.

id: Option<String>items: Option<JSONSchemaPropsOrArray>max_items: Option<i64>max_length: Option<i64>max_properties: Option<i64>maximum: Option<f64>min_items: Option<i64>min_length: Option<i64>min_properties: Option<i64>minimum: Option<f64>multiple_of: Option<f64>not: Option<Box<JSONSchemaProps>>nullable: Option<bool>one_of: Option<Vec<JSONSchemaProps>>pattern: Option<String>pattern_properties: Option<BTreeMap<String, JSONSchemaProps>>properties: Option<BTreeMap<String, JSONSchemaProps>>required: Option<Vec<String>>title: Option<String>type_: Option<String>unique_items: Option<bool>x_kubernetes_embedded_resource: Option<bool>

x-kubernetes-embedded-resource defines that the value is an embedded Kubernetes runtime.Object, with TypeMeta and ObjectMeta. The type must be object. It is allowed to further restrict the embedded object. kind, apiVersion and metadata are validated automatically. x-kubernetes-preserve-unknown-fields is allowed to be true, but does not have to be if the object is fully specified (up to kind, apiVersion, metadata).

x_kubernetes_int_or_string: Option<bool>

x-kubernetes-int-or-string specifies that this value is either an integer or a string. If this is true, an empty type is allowed and type as child of anyOf is permitted if following one of the following patterns:

anyOf:

type: integer

type: string

allOf:

anyOf:

type: integer

type: string

... zero or more

x_kubernetes_list_map_keys: Option<Vec<String>>

x-kubernetes-list-map-keys annotates an array with the x-kubernetes-list-type map by specifying the keys used as the index of the map.

This tag MUST only be used on lists that have the "x-kubernetes-list-type" extension set to "map". Also, the values specified for this attribute must be a scalar typed field of the child structure (no nesting is supported).

The properties specified must either be required or have a default value, to ensure those properties are present for all list items.

x_kubernetes_list_type: Option<String>

x-kubernetes-list-type annotates an array to further describe its topology. This extension must only be used on lists and may have 3 possible values:

atomic: the list is treated as a single entity, like a scalar. Atomic lists will be entirely replaced when updated. This extension may be used on any type of list (struct, scalar, ...).

set: Sets are lists that must not have multiple items with the same value. Each value must be a scalar, an object with x-kubernetes-map-type atomic or an array with x-kubernetes-list-type atomic.

map: These lists are like maps in that their elements have a non-index key used to identify them. Order is preserved upon merge. The map tag must only be used on a list with elements of type object. Defaults to atomic for arrays.

x_kubernetes_map_type: Option<String>

x-kubernetes-map-type annotates an object to further describe its topology. This extension must only be used when type is object and may have 2 possible values:

granular: These maps are actual maps (key-value pairs) and each fields are independent from each other (they can each be manipulated by separate actors). This is the default behaviour for all maps.

atomic: the list is treated as a single entity, like a scalar. Atomic maps will be entirely replaced when updated.

x_kubernetes_preserve_unknown_fields: Option<bool>

x-kubernetes-preserve-unknown-fields stops the API server decoding step from pruning fields which are not specified in the validation schema. This affects fields recursively, but switches back to normal pruning behaviour if nested properties or additionalProperties are specified in the schema. This can either be true or undefined. False is forbidden.