k8s_openapi::merge_strategies

Module list

Strategies for merging std::vec::Vecs.

These correspond to JSONSchemaProps.x-kubernetes-list-type.

Functions

atomic

The whole list is treated as one scalar value, and will be replaced with the new (non-None) value.

map

The list is treated as a map.

set

The list is treated as a set.