Trait k8s_openapi::Metadata

pub trait Metadata: Resource {
    type Ty;
    fn metadata(&self) -> &<Self as Metadata>::Ty;

    fn metadata_mut(&mut self) -> &mut <Self as Metadata>::Ty;
}

A trait applied to all Kubernetes resources that have metadata.

Associated Types

type Ty

The type of the metadata object.

Required methods

fn metadata(&self) -> &<Self as Metadata>::Ty

Gets a reference to the metadata of this resource value.

fn metadata_mut(&mut self) -> &mut <Self as Metadata>::Ty

Gets a mutable reference to the metadata of this resource value.