Trait k8s_openapi::Metadata

pub trait Metadata: Resource {
    type Ty;
    fn metadata(&self) -> &Self::Ty;

    fn metadata_mut(&mut self) -> &mut Self::Ty;
}

A trait applied to all Kubernetes resources that have metadata.

Associated Types

type Ty

The type of the metadata object.

Loading content...
Required methods

fn metadata(&self) -> &Self::Ty

Gets a reference to the metadata of this resource value.

fn metadata_mut(&mut self) -> &mut Self::Ty

Gets a mutable reference to the metadata of this resource value.

Loading content...