Trait k8s_openapi::Resource

pub trait Resource {
    const API_VERSION: &'static str;
    const GROUP: &'static str;
    const KIND: &'static str;
    const VERSION: &'static str;
}

A trait applied to all Kubernetes resources.

Associated Constants

const API_VERSION: &'static str

The API version of the resource. This is a composite of Resource::GROUP and Resource::VERSION (eg "apiextensions.k8s.io/v1beta1") or just the version for resources without a group (eg "v1").

This is the string used in the apiVersion field of the resource's serialized form.

const GROUP: &'static str

The group of the resource, or the empty string if the resource doesn't have a group.

const KIND: &'static str

The kind of the resource.

This is the string used in the kind field of the resource's serialized form.

const VERSION: &'static str

The version of the resource.

Loading content...