Trait k8s_openapi::Resource

pub trait Resource {
    fn api_version() -> &'static str
    where
        Self: Sized;

    fn group() -> &'static str
    where
        Self: Sized;

    fn kind() -> &'static str
    where
        Self: Sized;

    fn version() -> &'static str
    where
        Self: Sized;
}

A trait applied to all Kubernetes resources.

Required methods

fn api_version() -> &'static str where
Self: Sized,

The API version of the resource. This is a composite of Resource::group and Resource::version (eg "apiextensions.k8s.io/v1beta1") or just the version for resources without a group (eg "v1").

This is the string used in the apiVersion field of the resource's serialized form.

fn group() -> &'static str where
Self: Sized,

The group of the resource, or the empty string if the resource doesn't have a group.

fn kind() -> &'static str where
Self: Sized,

The kind of the resource.

This is the string used in the kind field of the resource's serialized form.

fn version() -> &'static str where
Self: Sized,

The version of the resource.

Loading content...